% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_easy_example.R
\name{array_easy_example}
\alias{array_easy_example}
\title{Example array}
\usage{
array_easy_example(I = 2, J = 3, K = 4)
}
\arguments{
\item{I}{natural number less than 10}

\item{J}{natural number less than 10}

\item{K}{natural number less than 10}
}
\value{
An array of three dimension.
}
\description{
Make a three dim array whose component is its index. For example

 \code{a[2,3,4] = 234}
}
\examples{



   a <-array_easy_example(2,3,4)



}
