% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Author_vs_Chakraborty_for_AUC.R
\name{Author_vs_classic_for_AUC}
\alias{Author_vs_classic_for_AUC}
\title{validation of AUC calculation}
\usage{
Author_vs_classic_for_AUC(StanS4class)
}
\arguments{
\item{StanS4class}{An \R  object of class \emph{\code{ \link{stanfitExtended}}} inherited from the S4 class  \code{\link[rstan]{stanfit}}  that can be passed to the \code{\link{DrawCurves}()} and  \code{\link{p_value_of_the_Bayesian_sense_for_chi_square_goodness_of_fit}()}, ..., etc.}
}
\value{
AUCs
}
\description{
This is for the author.
}
