% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Chi_square_goodness_of_fit_in_case_of_MRMC_Posterior_Mean.R
\name{Chi_square_goodness_of_fit_in_case_of_MRMC_Posterior_Mean}
\alias{Chi_square_goodness_of_fit_in_case_of_MRMC_Posterior_Mean}
\title{Chi square in the case of MRMC.}
\usage{
Chi_square_goodness_of_fit_in_case_of_MRMC_Posterior_Mean(StanS4class,
  summary = TRUE)
}
\arguments{
\item{StanS4class}{An \R  object of class \emph{\code{ \link{stanfitExtended}}} inherited from the S4 class  \code{\link[rstan]{stanfit}}  that can be passed to the \code{\link{DrawCurves}()} and  \code{\link{p_value_of_the_Bayesian_sense_for_chi_square_goodness_of_fit}()}, ..., etc.}

\item{summary}{Logical: \code{TRUE} of \code{FALSE}. Whether to print the verbose summary, i.e., logical; If \code{TRUE} then verbose summary is printed in the \R console. If \code{FALSE}, the output is minimal. I regret, this variable name should be verbose.}
}
\value{
A list, contains \{ \eqn{\chi^2(Data|\theta_i)} ; i= 1,2,3,...N\}, where N is the numberMCMC samples.
}
\description{
Not a posteriror predective chi square, but the posterior mean.

Let \eqn{\chi(y|\theta)} be a chi square goodness of fit statistics of our hierarchical Bayesian Model for , i.e.,

\deqn{\chi(D|\theta) := \sum _{r=1}^R \sum _{m=1}^M \sum _{c=1}^C \biggr( \frac{[ H_{c,m,r}-N_L\times p_{c,m,r}]^2}{N_L\times p_{c,m,r}}+
                                                                 \frac{[F_{c,m,r}-(\lambda _{c} -\lambda _{c+1} )\times N_{L}]^2}{(\lambda _{c} -\lambda _{c+1} )\times N_{L} }\biggr).
}

where  a dataset \eqn{D} denotes \eqn{ (F_{c,m,r}, H_{c,m,r}) } and model parameter \eqn{\theta}.

Then the function calculates the quantities

     \deqn{\chi(D|\theta_1),\chi(D|\theta_2),\chi(D|\theta_3),   \cdots,  \chi(D|\theta_N)        }

Foe MCMC samples;



      \deqn{   \theta_1 ,   \theta_2 ,   \theta_3,    \cdots ,    \theta_N         }.




Also,  it evaluate the posterior mean of the chi square, statistc, that is,


      \deqn{\frac{1}{N} \sum _i ^N \chi(D|\theta_i)}

which is approximation of the following posterior integral;


      \deqn{\int   \chi(D|\theta)\pi(\theta|D) d\theta}

where \eqn{\pi(\theta|D)} denotes the posterior probability.
}
\author{
Issei Tsunoda
}
