% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearWorkspace.R
\name{clearWorkspace}
\alias{clearWorkspace}
\title{Clear Work Space}
\usage{
clearWorkspace()
}
\description{
If functions are masked in global environment, I use this.
this function has no variables.
}
\author{
Issei Tsunoda
}
