% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Close_all_graphic_devices.R
\name{Close_all_graphic_devices}
\alias{Close_all_graphic_devices}
\title{Close the Graphic Device}
\usage{
Close_all_graphic_devices()
}
\description{
Close the graphic device to avoid errors in R CMD check.
}
\examples{

\donttest{
 #    Open the graphic devices

   grDevices::dev.new();plot(stats::runif(100),stats::runif(100))
   grDevices::dev.new();plot(stats::runif(100),stats::runif(100))
   grDevices::dev.new();plot(stats::runif(100),stats::runif(100))
   grDevices::dev.new();plot(stats::runif(100),stats::runif(100))
   grDevices::dev.new();plot(stats::runif(100),stats::runif(100))
   grDevices::dev.new();plot(stats::runif(100),stats::runif(100))


#    Close the graphic device

       Close_all_graphic_devices()


#'}# dottest
}
