% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Color_Message.R
\name{color_message}
\alias{color_message}
\title{message with colored item}
\usage{
color_message(words, ..., type = 1, print_debug = TRUE)
}
\arguments{
\item{words}{Characters}

\item{...}{Characters}

\item{type}{An integer}

\item{print_debug}{A logical, whether prints a message}
}
\value{
NULL or print
}
\description{
message with colored item
}
\examples{

 color_message("aaaaa","bbbbb",type = 2,print_debug = TRUE)

}
