% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelComparison.R
\name{compare}
\alias{compare}
\title{model comparison}
\usage{
compare(NI, ite = 1111)
}
\arguments{
\item{NI}{images}

\item{ite}{iteration
####1#### ####2#### ####3#### ####4#### ####5#### ####6#### ####7#### ####8#### ####9####}
}
\description{
This is a model comparison.
}
