% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelComparison.R
\name{comparison}
\alias{comparison}
\title{model comparison}
\usage{
comparison(Number.of.variation.of.NL, Number.of.images, ite = 1111,
  DrawCurve = FALSE, dig = 3, e = 0)
}
\arguments{
\item{Number.of.variation.of.NL}{Lesion}

\item{Number.of.images}{images}

\item{ite}{iteration}

\item{DrawCurve}{This is a dichotomous, i.e., TRUE or FALSE. If you want to draw the FROC and AFROC curves, then you set \code{DrawCurve =TRUE}, if not then \code{DrawCurve =FALSE}.
The reason why the author make this variable \code{DrawCurve} is that it takes long time to draw curves, and thus default value is \code{FALSE}.}

\item{dig}{An argument of \code{rstan::stan()} in which it is named \code{iter}.  It means the Significant digits, used in stan Cancellation.
default = 5,}

\item{e}{exp for false alarms}
}
\description{
This is a model comparison.
}
