% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_all_models.R
\name{compile_all_models_in_pkg_BayesianFROC}
\alias{compile_all_models_in_pkg_BayesianFROC}
\title{Compile all stanfiles in pkg BayesianFROC}
\usage{
compile_all_models_in_pkg_BayesianFROC()
}
\value{
none
}
\description{
Compile all stanfiles in pkg BayesianFROC
}
\examples{
\dontrun{
# compile_all_models_in_pkg_BayesianFROC()
}

}
