% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset.R
\name{create_dataset}
\alias{create_dataset}
\title{Create a dataset}
\usage{
create_dataset()
}
\value{
A return value, i.e. a list, which are used to build FROC models.
}
\description{
Create the Passing data to  the  function \code{ \link{fit_Bayesian_FROC} }.


   This is an interactive creator of dataset for FROC data. Using this return value, you can build the FROC model for your data by applying the  function \code{ \link{fit_Bayesian_FROC}} in this package.
}
