% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_dataset_srsc.R
\docType{data}
\name{d}
\alias{d}
\title{Data: A Single Reader and A Single Modality}
\description{
A list, representing FROC data. This is used to build a hierarchical FROC model. This data is exactly same as dataList.Chakra.1.
}
\details{
This data is same as \code{\link{dataList.Chakra.1.with.explantation}}.
 The author name it  \code{d} for the sake of simplicity, that is, it is easy to write, because only one character!!
}
\references{
Maximum likelihood analysis of free-response  receiver operating characteristic (FROC) data, Dev P. Chakraborty.
}
\seealso{
\code{\link{dataList.Chakra.1.with.explantation}} which is exactly same in this data \code{d}.
}
\author{
Issei Tsunoda \email{tsunoda.issei1111@gmail.com }
}
