% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dark_theme.R
\name{dark_theme}
\alias{dark_theme}
\title{Dark Theme}
\usage{
dark_theme(type = 1)
}
\arguments{
\item{type}{An integer}
}
\value{
Nothing
}
\description{
Executing this function before plotting,
the plot region becomes the dark theme.
}
\details{
A function specifies the color
in graphic devices.
}
\examples{
 \donttest{

 dark_theme(1)

 graphics::plot(c(1,2,3),c(1,2,3))


 dark_theme(2)

 graphics::plot(c(1,2,3),c(1,2,3))


 # 2019.05.21 Revised.




 dark_theme(3)

 graphics::plot(c(1,2,3),c(1,2,3))


 dark_theme(4)

 graphics::plot(c(1,2,3),c(1,2,3))



#  2019 Oct 19 Revised

}# dottest
}
