% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_dataset_MRMC.R
\docType{data}
\name{data_of_36_readers_and_a_single_modality}
\alias{data_of_36_readers_and_a_single_modality}
\title{36 readers and a sinle modality data}
\description{
Big data example in the FROC context
}
\details{
Frequentist methods fails when a sample size is large.
Namely, p value  monotonically decreases when the sample size tends to large.

On the other hands, in Bayesian methods, the large samples such as large readers in FROC context
fails the MCMC algorithm. Thus Bayesian methods is also not free from such large sample problem in this sense.

\strong{  This dataset is made for validation that wheter Bayes factor well work } \emph{ which is a subset of data \code{dataList.Chakra.Web.orderd} }
\describe{
\item{ the number of modalities, denoted by \code{M}   }{  1 modality  }
\item{ the number of Confidences, denoted by \code{C}   }{  5 Confidence levels }
\item{ the number of readers, denoted by \code{Q}  }{  36 readers}
}



\strong{Contents of \code{data_of_36_readers_and_a_single_modality}}


 \code{NL}  = 142 (Number of Lesions)

 \code{NI} = 57 (Number of Images)#'





\strong{\emph{ Contents:  }}

 \emph{          Multiple readers and multiple modalities case, i.e., MRMC case   }

\tabular{ccccc}{
 \strong{ModalityID } \tab   \strong{ReaderID }  \tab  \strong{ Confidence levels} \tab   \strong{No. of false alarms} \tab   \strong{No. of hits}.\cr
  \code{m} \tab  \code{ q}  \tab   \code{c} \tab  \code{ f} \tab \code{ h}\cr
  -------------- \tab ------------- \tab ------------------------ \tab  ------------------- \tab ----------------\cr
1 \tab  1 \tab 5 \tab  0 \tab 12\cr
1 \tab  1 \tab 4 \tab  3 \tab 22\cr
1 \tab  1 \tab 3 \tab  7 \tab 18\cr
1 \tab  1 \tab 2 \tab 12 \tab 18\cr
1 \tab  1 \tab 1 \tab  8 \tab 15\cr
1 \tab  2 \tab 5 \tab  0 \tab 14\cr
1 \tab  2 \tab 4 \tab  4 \tab 24\cr
1 \tab  2 \tab 3 \tab  9 \tab 17\cr
1 \tab  2 \tab 2 \tab 14 \tab 15\cr
1 \tab  2 \tab 1 \tab 10 \tab  6\cr
1 \tab  3 \tab 5 \tab  0 \tab 26\cr
1 \tab  3 \tab 4 \tab  3 \tab 39\cr
1 \tab  3 \tab 3 \tab  6 \tab 23\cr
1 \tab  3 \tab 2 \tab 11 \tab 16\cr
1 \tab  3 \tab 1 \tab  7 \tab  6\cr
1 \tab  4 \tab 5 \tab  0 \tab  9\cr
1 \tab  4 \tab 4 \tab  1 \tab 17\cr
1 \tab  4 \tab 3 \tab  4 \tab 15\cr
1 \tab  4 \tab 2 \tab  8 \tab 18\cr
1 \tab  4 \tab 1 \tab  5 \tab 25\cr
1 \tab  5 \tab 5 \tab  0 \tab  9\cr
1 \tab  5 \tab 4 \tab  2 \tab 17\cr
1 \tab  5 \tab 3 \tab  5 \tab 16\cr
1 \tab  5 \tab 2 \tab  9 \tab 19\cr
1 \tab  5 \tab 1 \tab  6 \tab 27\cr
1 \tab  6 \tab 5 \tab  0 \tab 39\cr
1 \tab  6 \tab 4 \tab  0 \tab 46\cr
1 \tab  6 \tab 3 \tab  2 \tab 22\cr
1 \tab  6 \tab 2 \tab 15 \tab 13\cr
1 \tab  6 \tab 1 \tab  2 \tab  3\cr
1 \tab  7 \tab 5 \tab  0 \tab  9\cr
1 \tab  7 \tab 4 \tab  1 \tab 17\cr
1 \tab  7 \tab 3 \tab  4 \tab 14\cr
1 \tab  7 \tab 2 \tab  8 \tab 16\cr
1 \tab  7 \tab 1 \tab  5 \tab 17\cr
1 \tab  8 \tab 5 \tab  1 \tab 11\cr
1 \tab  8 \tab 4 \tab  5 \tab 19\cr
1 \tab  8 \tab 3 \tab 10 \tab 16\cr
1 \tab  8 \tab 2 \tab 16 \tab 17\cr
1 \tab  8 \tab 1 \tab 12 \tab 15\cr
1 \tab  9 \tab 5 \tab  0 \tab 15\cr
1 \tab  9 \tab 4 \tab  1 \tab 26\cr
1 \tab  9 \tab 3 \tab  3 \tab 20\cr
1 \tab  9 \tab 2 \tab  6 \tab 18\cr
1 \tab  9 \tab 1 \tab  4 \tab 12\cr
1 \tab 10 \tab 5 \tab  0 \tab 31\cr
1 \tab 10 \tab 4 \tab  4 \tab 40\cr
1 \tab 10 \tab 3 \tab  8 \tab 22\cr
1 \tab 10 \tab 2 \tab 13 \tab 16\cr
1 \tab 10 \tab 1 \tab  9 \tab  5\cr
1 \tab 11 \tab 5 \tab  0 \tab 13\cr
1 \tab 11 \tab 4 \tab  2 \tab 23\cr
1 \tab 11 \tab 3 \tab  5 \tab 19\cr
1 \tab 11 \tab 2 \tab  9 \tab 19\cr
1 \tab 11 \tab 1 \tab  6 \tab 17\cr
1 \tab 12 \tab 5 \tab  0 \tab  8\cr
1 \tab 12 \tab 4 \tab  3 \tab 16\cr
1 \tab 12 \tab 3 \tab  7 \tab 15\cr
1 \tab 12 \tab 2 \tab 11 \tab 17\cr
1 \tab 12 \tab 1 \tab  8 \tab 22\cr
1 \tab 13 \tab 5 \tab  0 \tab 13\cr
1 \tab 13 \tab 4 \tab  1 \tab 23\cr
1 \tab 13 \tab 3 \tab  4 \tab 19\cr
1 \tab 13 \tab 2 \tab  7 \tab 21\cr
1 \tab 13 \tab 1 \tab  4 \tab 20\cr
1 \tab 14 \tab 5 \tab  0 \tab 36\cr
1 \tab 14 \tab 4 \tab  4 \tab 45\cr
1 \tab 14 \tab 3 \tab  9 \tab 22\cr
1 \tab 14 \tab 2 \tab 14 \tab 13\cr
1 \tab 14 \tab 1 \tab 10 \tab  3\cr
1 \tab 15 \tab 5 \tab  0 \tab 17\cr
1 \tab 15 \tab 4 \tab  2 \tab 27\cr
1 \tab 15 \tab 3 \tab  5 \tab 20\cr
1 \tab 15 \tab 2 \tab  9 \tab 18\cr
1 \tab 15 \tab 1 \tab  6 \tab 10\cr
1 \tab 16 \tab 5 \tab  0 \tab  8\cr
1 \tab 16 \tab 4 \tab  4 \tab 15\cr
1 \tab 16 \tab 3 \tab  8 \tab 13\cr
1 \tab 16 \tab 2 \tab 13 \tab 16\cr
1 \tab 16 \tab 1 \tab  9 \tab 22\cr
1 \tab 17 \tab 5 \tab  0 \tab  9\cr
1 \tab 17 \tab 4 \tab  1 \tab 16\cr
1 \tab 17 \tab 3 \tab  4 \tab 15\cr
1 \tab 17 \tab 2 \tab  8 \tab 17\cr
1 \tab 17 \tab 1 \tab  5 \tab 20\cr
1 \tab 18 \tab 5 \tab  0 \tab 12\cr
1 \tab 18 \tab 4 \tab  2 \tab 21\cr
1 \tab 18 \tab 3 \tab  6 \tab 17\cr
1 \tab 18 \tab 2 \tab 10 \tab 17\cr
1 \tab 18 \tab 1 \tab  7 \tab 12\cr
1 \tab 19 \tab 5 \tab  0 \tab 19\cr
1 \tab 19 \tab 4 \tab  3 \tab 33\cr
1 \tab 19 \tab 3 \tab  8 \tab 21\cr
1 \tab 19 \tab 2 \tab 12 \tab 19\cr
1 \tab 19 \tab 1 \tab  9 \tab 13\cr
1 \tab 20 \tab 5 \tab  0 \tab  8\cr
1 \tab 20 \tab 4 \tab  1 \tab 15\cr
1 \tab 20 \tab 3 \tab  3 \tab 14\cr
1 \tab 20 \tab 2 \tab  6 \tab 16\cr
1 \tab 20 \tab 1 \tab  4 \tab 21\cr
1 \tab 21 \tab 5 \tab  0 \tab 33\cr
1 \tab 21 \tab 4 \tab  2 \tab 41\cr
1 \tab 21 \tab 3 \tab  5 \tab 21\cr
1 \tab 21 \tab 2 \tab  9 \tab 13\cr
1 \tab 21 \tab 1 \tab  6 \tab  3\cr
1 \tab 22 \tab 5 \tab  0 \tab 15\cr
1 \tab 22 \tab 4 \tab  3 \tab 26\cr
1 \tab 22 \tab 3 \tab  7 \tab 20\cr
1 \tab 22 \tab 2 \tab 12 \tab 20\cr
1 \tab 22 \tab 1 \tab  8 \tab 15\cr
1 \tab 23 \tab 5 \tab  0 \tab  9\cr
1 \tab 23 \tab 4 \tab  4 \tab 17\cr
1 \tab 23 \tab 3 \tab  8 \tab 15\cr
1 \tab 23 \tab 2 \tab 12 \tab 18\cr
1 \tab 23 \tab 1 \tab  9 \tab 23\cr
1 \tab 24 \tab 5 \tab  0 \tab 10\cr
1 \tab 24 \tab 4 \tab  0 \tab 19\cr
1 \tab 24 \tab 3 \tab  3 \tab 17\cr
1 \tab 24 \tab 2 \tab  6 \tab 20\cr
1 \tab 24 \tab 1 \tab  4 \tab 23\cr
1 \tab 25 \tab 5 \tab  0 \tab  8\cr
1 \tab 25 \tab 4 \tab  1 \tab 15\cr
1 \tab 25 \tab 3 \tab  3 \tab 14\cr
1 \tab 25 \tab 2 \tab  6 \tab 17\cr
1 \tab 25 \tab 1 \tab  4 \tab 22\cr
1 \tab 26 \tab 5 \tab  0 \tab 12\cr
1 \tab 26 \tab 4 \tab  1 \tab 21\cr
1 \tab 26 \tab 3 \tab  4 \tab 18\cr
1 \tab 26 \tab 2 \tab  8 \tab 19\cr
1 \tab 26 \tab 1 \tab  5 \tab 18\cr
1 \tab 27 \tab 5 \tab  0 \tab 19\cr
1 \tab 27 \tab 4 \tab  1 \tab 32\cr
1 \tab 27 \tab 3 \tab  4 \tab 18\cr
1 \tab 27 \tab 2 \tab  7 \tab 13\cr
1 \tab 27 \tab 1 \tab  5 \tab  4\cr
1 \tab 28 \tab 5 \tab  1 \tab 10\cr
1 \tab 28 \tab 4 \tab  5 \tab 18\cr
1 \tab 28 \tab 3 \tab  9 \tab 16\cr
1 \tab 28 \tab 2 \tab 15 \tab 19\cr
1 \tab 28 \tab 1 \tab 11 \tab 26\cr
1 \tab 29 \tab 5 \tab  0 \tab 16\cr
1 \tab 29 \tab 4 \tab  2 \tab 27\cr
1 \tab 29 \tab 3 \tab  6 \tab 21\cr
1 \tab 29 \tab 2 \tab 10 \tab 20\cr
1 \tab 29 \tab 1 \tab  7 \tab 16\cr
1 \tab 30 \tab 5 \tab  1 \tab  9\cr
1 \tab 30 \tab 4 \tab  4 \tab 18\cr
1 \tab 30 \tab 3 \tab  9 \tab 16\cr
1 \tab 30 \tab 2 \tab 14 \tab 19\cr
1 \tab 30 \tab 1 \tab 10 \tab 25\cr
1 \tab 31 \tab 5 \tab  0 \tab 10\cr
1 \tab 31 \tab 4 \tab  3 \tab 19\cr
1 \tab 31 \tab 3 \tab  7 \tab 16\cr
1 \tab 31 \tab 2 \tab 11 \tab 18\cr
1 \tab 31 \tab 1 \tab  8 \tab 20\cr
1 \tab 32 \tab 5 \tab  1 \tab 12\cr
1 \tab 32 \tab 4 \tab  5 \tab 22\cr
1 \tab 32 \tab 3 \tab 10 \tab 18\cr
1 \tab 32 \tab 2 \tab 15 \tab 19\cr
1 \tab 32 \tab 1 \tab 11 \tab 18\cr
1 \tab 33 \tab 5 \tab  1 \tab 14\cr
1 \tab 33 \tab 4 \tab  6 \tab 24\cr
1 \tab 33 \tab 3 \tab 11 \tab 18\cr
1 \tab 33 \tab 2 \tab 16 \tab 17\cr
1 \tab 33 \tab 1 \tab 12 \tab 10\cr
1 \tab 34 \tab 5 \tab  0 \tab 34\cr
1 \tab 34 \tab 4 \tab  3 \tab 43\cr
1 \tab 34 \tab 3 \tab  8 \tab 22\cr
1 \tab 34 \tab 2 \tab 12 \tab 14\cr
1 \tab 34 \tab 1 \tab  9 \tab  3\cr
1 \tab 35 \tab 5 \tab  0 \tab  9\cr
1 \tab 35 \tab 4 \tab  1 \tab 17\cr
1 \tab 35 \tab 3 \tab  4 \tab 15\cr
1 \tab 35 \tab 2 \tab  8 \tab 18\cr
1 \tab 35 \tab 1 \tab  5 \tab 25\cr
1 \tab 36 \tab 5 \tab  1 \tab 17\cr
1 \tab 36 \tab 4 \tab  6 \tab 31\cr
1 \tab 36 \tab 3 \tab 11 \tab 20\cr
1 \tab 36 \tab 2 \tab 16 \tab 17\cr
1 \tab 36 \tab 1 \tab 12 \tab  9
}
}
\examples{


#----------------------------------------------------------------------------------------
#                        Show data by table
#----------------------------------------------------------------------------------------



                        viewdata(data_of_36_readers_and_a_single_modality)


plot_FPF_and_TPF_from_a_dataset(data_of_36_readers_and_a_single_modality)

####1#### ####2#### ####3#### ####4#### ####5#### ####6#### ####7#### ####8#### ####9####
#----------------------------------------------------------------------------------------
#                       make this data from functions in this package
#----------------------------------------------------------------------------------------



v  <- v_truth_creator_for_many_readers_MRMC_data(M=1,Q=36)
m  <- mu_truth_creator_for_many_readers_MRMC_data(M=1,Q=36)
d  <- create_dataList_MRMC(mu.truth = m,v.truth = v)


# The last object named d is the desired dataset.

\donttest{
 fit_Bayesian_FROC(data_of_36_readers_and_a_single_modality,ite = 1111)
}

}
\references{
Example data of Jafroc software
}
\seealso{
Not \code{\link{dataList.Chakra.Web} }
But \code{\link{dataList.Chakra.Web.orderd} }
 Not \code{\link{dd} }
}
\author{
Issei Tsunoda \email{tsunoda.issei1111@gmail.com }
}
