% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_dataset.R
\docType{data}
\name{dataList.Chakra.Web.orderd}
\alias{dataList.Chakra.Web.orderd}
\title{Multiple reader and Multiple modality data}
\description{
This is used to build a hierarchical FROC model.
}
\details{
This data appeared in Chakraborty's JAFROC. I have  ordered so that the modality ID means the  order of AUC. For example modality ID = 1 means its AUC is the highest. modalityID = 2 means its AUC is the second.
}
\references{
Maximum likelihood analysis of free-response  receiver operating characteristic (FROC) data, Dev P. Chakraborty.
}
\author{
Issei Tsunoda \email{tsunoda.issei1111@gmail.com }
}
\keyword{FROC}
\keyword{Non-hierarchical}
\keyword{Single}
\keyword{and}
\keyword{data.}
\keyword{modality}
\keyword{reader}
