% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_dataset_srsc.R
\docType{data}
\name{dataList.High}
\alias{dataList.High}
\title{Data: Single reader and Single modality}
\description{
A list, representing FROC data. This is used to build a hierarchical FROC model. This data is same as dataList.Chakra.1.
}
\details{
This data-set is fictitious.
}
\references{
Maximum likelihood analysis of free-response  receiver operating characteristic (FROC) data, Dev P. Chakraborty.
}
\seealso{
\code{\link{dataList.Chakra.1.with.explantation}}
}
\author{
Issei Tsunoda \email{tsunoda.issei1111@gmail.com }
}
