% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_dataset_MRMC.R
\docType{data}
\name{dataList.one.modality}
\alias{dataList.one.modality}
\title{Multiple reader and one modality data for fit_MRMC_versionTWO}
\description{
This is used to build a hierarchical FROC model.
}
\details{
This data contains only one modality. If see = 12, then the model has converged.
}
\references{
Nothing in 2018
}
\author{
Issei Tsunoda \email{tsunoda.issei1111@gmail.com }
}
