% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_creator_for_many_Readers.R
\name{dataset_creator_for_many_Readers}
\alias{dataset_creator_for_many_Readers}
\title{create data for MRMC}
\usage{
dataset_creator_for_many_Readers(M, Q)
}
\arguments{
\item{M}{a positive integer, modality}

\item{Q}{a positive integer, reader}
}
\value{
data, to which fit a model
}
\description{
create data for MRMC
}
\examples{

d <- dataset_creator_for_many_Readers(1,11)


}
