% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_creator_new_version.R
\name{dataset_creator_new_version}
\alias{dataset_creator_new_version}
\title{Create a dataset (version 2)}
\usage{
dataset_creator_new_version()
}
\value{
A return value, i.e. a list, which are used to build FROC models.
}
\description{
This is an interactive creator of dataset for FROC data. Using this return value, you can build the FROC model for your data by applying the other functions in this package.
}
