% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_dataset_MRMC.R
\docType{data}
\name{ddddddd}
\alias{ddddddd}
\title{Multiple reader and 2 modalities data such that all modalities have same AUC.}
\description{
This is a subset of \code{ \link{dataList.Chakra.Web.orderd}}
}
\details{
The author made this dataset to validate the scheme of Bayes factor well works in our Bayesian FROC models

\strong{  This dataset is made for validation that wheter Bayes factor well work } \emph{ which is a subset of data \code{dataList.Chakra.Web.orderd} }

\describe{
\item{ dddddd$M   }{  2 modalities of almost \strong{\emph{same}} AUC  }
\item{ dddddd$C   }{  3 Confidence levels }
\item{ dddddd$Q   }{  2 readers}
}

If Bayes factor admit the null hypothesis that all modality are same, that is, 1-st and 2-nd modality of  \code{ \link{dataList.Chakra.Web.orderd}} are same,
then, the Bayes factor well works.





\strong{Contents of dddddd}


 \code{NL}  = 142 (Number of Lesions)

 \code{NI} = 199 (Number of Images)#'





\strong{\emph{ Contents:  }}

 \emph{          Multiple readers and multiple modalities case, i.e., MRMC case   }
}
\examples{


#========================================================================================
#                        Show data by table
#========================================================================================



                        viewdata(ddddddd)




####1#### ####2#### ####3#### ####4#### ####5#### ####6#### ####7#### ####8#### ####9####
#========================================================================================
#                       make an object dddd from an object dataList.Chakra.Web.orderd
#========================================================================================


ddd  <-  data.frame(m=dataList.Chakra.Web.orderd$m,
                    q=dataList.Chakra.Web.orderd$q,
                    c=dataList.Chakra.Web.orderd$c,
                    h=dataList.Chakra.Web.orderd$h,
                    f=dataList.Chakra.Web.orderd$f
)

dddd <- ddd[ddd$q < 3,]

# The following code extract the first and the second modality from dd
dddd <- dddd[dddd$m < 3,]  #  Reduce the dataset ddd, i.e., dd
dddd <- dddd[dddd$c <4,]
ddd <- list(
  m=dddd$m,
  q=dddd$q,
  c=dddd$c,
  h=dddd$h,
  f=dddd$f,
  NL=142,
  NI=199, # 2020 April 6
  C=max(dddd$c),
  M=max(dddd$m),
  Q=max(dddd$q)
)

ddddddd <-ddd


# This dataset is made in 2019 July 6, for the aim of easy exihibition
# This dataset is very minimum, and it is easy to view


}
\references{
Example data of Jafroc software
}
\seealso{
Not \code{\link{dataList.Chakra.Web} }
But \code{\link{dataList.Chakra.Web.orderd} }
 Not \code{\link{dd} }
}
\author{
Issei Tsunoda \email{tsunoda.issei1111@gmail.com }
}
