% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_Bayesian_FROC.R
\name{demo_Bayesian_FROC}
\alias{demo_Bayesian_FROC}
\title{demonstration}
\usage{
demo_Bayesian_FROC()
}
\value{
none
}
\description{
demonstration
}
\examples{
 \donttest{


 demo_Bayesian_FROC()


}# dottest
}
