% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_Bayesian_FROC_without_pause.R
\name{demo_Bayesian_FROC_without_pause}
\alias{demo_Bayesian_FROC_without_pause}
\title{demonstration without pausing}
\usage{
demo_Bayesian_FROC_without_pause()
}
\value{
none
}
\description{
demonstration without pausing. The author does not want to be bothered to hit Enter key. So,,,, made this. But now, I completely forget what codes run,,,,now 2020 Jul.
}
\examples{
 \dontrun{

demo_Bayesian_FROC_without_pause()



     Close_all_graphic_devices() # 2020 August
}

}
