% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFROC.R
\name{Draw_AUC}
\alias{Draw_AUC}
\title{draw AUC of AFROC}
\usage{
Draw_AUC(a = 0.13, b = 0.19, mesh.for.drawing.curve = 2222)
}
\arguments{
\item{a}{a generated parameter of model which characterize AFROC curve}

\item{b}{a generated parameter of model which characterize AFROC curve}

\item{mesh.for.drawing.curve}{A positive large integer, indicating number of dots drawing the curves, default =10000.}
}
\value{
none.
}
\description{
draw AUC of AFROC
}
\examples{



        Draw_AUC()


}
