% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawCurves.R
\name{DrawCurves}
\alias{DrawCurves}
\title{Draw the FROC  curves}
\usage{
DrawCurves(StanS4class, modalityID, readerID, title = TRUE,
  indexCFPCTP = FALSE, upper_x, upper_y, new.imaging.device = TRUE,
  Colour = TRUE, DrawFROCcurve = TRUE, DrawAFROCcurve = FALSE,
  DrawCFPCTP = TRUE, Draw.Flexible.upper_y = TRUE,
  Draw.Flexible.lower_y = TRUE)
}
\arguments{
\item{StanS4class}{An \R  object of class \emph{\code{ \link{stanfitExtended}}} inherited from the S4 class  \code{\link[rstan]{stanfit}}  that can be passed to the \code{\link{DrawCurves}()} and  \code{\link{p_value_of_the_Bayesian_sense_for_chi_square_goodness_of_fit}()}, ..., etc.}

\item{modalityID}{A positive integer vector indicating modalityID. If it is not given, then the first modality is chosen.}

\item{readerID}{A positive integer vector indicating  readerID. If it is not given, then the first reader is chosen.}

\item{title}{Logical: \code{TRUE} of \code{FALSE}.  If \code{TRUE}  (default), then title of curves are drawn.}

\item{indexCFPCTP}{TRUE of FALSE. If TRUE, then the cumulative false and hits are specified with its confidence level.}

\item{upper_x}{This is a upper bound for the axis of the horisontal coordinate of FROC curve.}

\item{upper_y}{This is a upper bound for the axis of the vertical coordinate of FROC curve.}

\item{new.imaging.device}{Logical: \code{TRUE} of \code{FALSE}. If TRUE (default), then open a new device to draw curve.
Using this we can draw curves in same plain by new.imaging.device=FALSE.}

\item{Colour}{Logical: \code{TRUE} of \code{FALSE}.  whether Colour of curves is dark theme or not.}

\item{DrawFROCcurve}{Logical: \code{TRUE} of \code{FALSE}.  Whether or not FROC curves are shown.}

\item{DrawAFROCcurve}{Logical: \code{TRUE} of \code{FALSE}.  Whether or not AFROC curves are shown.}

\item{DrawCFPCTP}{Logical: \code{TRUE} of \code{FALSE}.  Whether or not the pairs of \emph{FPF} and \emph{TPF} are shown.}

\item{Draw.Flexible.upper_y}{Logical: \code{TRUE} of \code{FALSE}.  Whether or not the upper bounds of vertical axis are determined automatically.}

\item{Draw.Flexible.lower_y}{Logical: \code{TRUE} of \code{FALSE}.  Whether or not the lower bounds of vertical axis are determined automatically.}
}
\description{
The function makes a plot of the FROC curve,  the AFROC curve and  \emph{FPF} and \emph{TPF}.
}
\details{
The function makes a plot of the FROC curves and AFROC curves for user's specified modality and user's specified reader.
Using this function \strong{repeatedly}, we can draw the different reader and modality in a  \strong{same} plane simultaneously.
So, we can visualize the difference of modality (reader).
}
\examples{
#================The first example======================================
\donttest{
#1) Build the S4 class object by the following:



  fit <- fit_Bayesian_FROC(
  BayesianFROC::dataList.Chakra.Web,    # data to which fit the model
                           ite=1111     # iteration of MCMC, very few, should be more.
                           )


# The object "fit" is an S4 class object
# whose S4 class is an inherited class from stanfit in the rstan package.

#<<Minor comments>>
#Note that return value "fit" is not an stanfit S4 object generated by rstan::stan(),
#but some inherited S4 class object which is an S4 object of
# some inherited S4 class from the stanfit class.





#2) Now, we obtain the S4 class object named "fit".
  # Using this S4 class object, we draw the curves by:



          DrawCurves(fit,modality = 1,reader = 4)

#From this code, FROC curve for the first modality and fourth reader is drawn.







#3) By changing, e.g., the modality,
   #we can draw the curves for different  modalities.
   #This shows the comparison of modalites.
   # In the following R script,
   #  the first draw curve for the 2 nd modality and the fourth reader,
   # and the second R script draw for the 3rd modality and the 4 th reader,
   # respectively.



            DrawCurves(fit,modality = 2,reader = 4)
            DrawCurves(fit,modality = 3,reader = 4)



# Curves are overwritten for the sake of comparison.
# When comparing modalities fitted by the hierarchical Bayesian Model to the same data,
# the upper FROC curve or AFROC curve, the better the AUC.









#4) By applying the function with respect to different modalities
#   in this manner, we can draw  AFROC (FROC) curves in the same plain.



#5) If you want to draw the FROC curves
#for reader ID =1,2,3,4 and modality ID =1,2, then the code is as follows;

                  DrawCurves(
                            fit,
                            modalityID = c(1,2,3,4),
                            readerID = c(1,2)
                            )
# Each color of curves corresponds to the modality ID.
# So, even if curves are different readers and same modality, then color is same.







#6) If you want to see only data points, then by DrawFROCcurve = F, it will be done.

DrawCurves(fit,
           DrawCFPCTP = TRUE,      # This implies data points are ploted.
           DrawFROCcurve = FALSE,  # From this, the curves are not drawn.
           modalityID = c(1,2,3,4),
           readerID  = c(1)
           )







#7) If you use the plot in submission and it is not allowed to use color, then
#   by Colour =F, you can get black and white plots, e.g.,


DrawCurves(fit,
          DrawCFPCTP = TRUE,
          DrawFROCcurve = TRUE,
          modalityID = c(1,2,3,4),
          readerID  = c(1),
          Colour = FALSE    # From this, you can get plots without colors.
          )







#'#8)  For AFROC, use DrawAFROCcurve = T

DrawCurves(fit,
          DrawFROCcurve = FALSE,
          DrawAFROCcurve = TRUE,
          modalityID = c(1,2,3,4),
          readerID  = c(1))







#================The Second Example======================================

#This function is available in the case of single reader and single modality.
#The reason why the maintainer separate the fitting and drawing curves is, in MRMC case,
#It tooks a time to drawing, but in the single reader and single modality case, drawing
# the curve is very fast, so in fitting process the curves are also depicted, however
# by this function user can draw the FROC curves.





#First, we prepare the data endowed with this package.





                          dat  <- get(data("dataList.Chakra.1"))




#Second, we run the stan funtion
#with data named "dat"  and the author's Bayesian model.





                             fit <-  fit_srsc_per_lesion(dat)





# Drawing the curves by


                             DrawCurves(fit)









#      Close the graphic device to avoid errors in R CMD check.

         Close_all_graphic_devices()
}# dottest
}
