% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawCurves_MRMC.R
\name{DrawCurves_MRMC}
\alias{DrawCurves_MRMC}
\title{Draw the FROC  curves for all modalities and readers}
\usage{
DrawCurves_MRMC(StanS4class)
}
\arguments{
\item{StanS4class}{This is an output of \code{rstan::stan()}.}
}
\description{
Draw the FROC  curves and AFROC curves for all modalities and readers, so very confused plots are drawn.
}
\examples{

\donttest{
  fit <- fit_Bayesian_FROC(
                           dataList.Chakra.Web.orderd,
                           ite = 1111,
                           summary =FALSE
                           )

               DrawCurves_MRMC(fit)

}# dottest

}
