% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawCurves_MRMC_pairwise_col.R
\name{DrawCurves_MRMC_pairwise_col}
\alias{DrawCurves_MRMC_pairwise_col}
\title{Draw the FROC  curves with Colour}
\usage{
DrawCurves_MRMC_pairwise_col(StanS4class, modalityID, readerID,
  new.imaging.device = TRUE, DrawFROCcurve = TRUE,
  DrawAFROCcurve = FALSE, DrawCFPCTP = TRUE,
  Draw.Flexible.upper_y = TRUE, Draw.Flexible.lower_y = TRUE)
}
\arguments{
\item{StanS4class}{This is an \R  object of class \emph{\code{ \link{stanfitExtended}}} inherited from the S4 class  \strong{\emph{\code{\link[rstan]{stanfit}}}}.}

\item{modalityID}{This is a vector indicating modalityID whose component is natural namber.}

\item{readerID}{This is a vector indicating readerID whose component is natural namber.}

\item{new.imaging.device}{Logical: \code{TRUE} of \code{FALSE}. If TRUE (default), then open a new device to draw curve.
Using this we can draw curves in same plain by new.imaging.device=FALSE.}

\item{DrawFROCcurve}{Logical: \code{TRUE} of \code{FALSE}. Whether  the FROC curve  is to be drawn.}

\item{DrawAFROCcurve}{Logical: \code{TRUE} of \code{FALSE}. Whether the  AFROC curve is to be drawn.}

\item{DrawCFPCTP}{Logical: \code{TRUE} of \code{FALSE}. Whether  the CFPCTP points are  to be drawn.}

\item{Draw.Flexible.upper_y}{Logical, that is \code{TRUE} or \code{FALSE}. Whether or not the upper bounds of vertical axis are determined automatically.}

\item{Draw.Flexible.lower_y}{Logical, that is \code{TRUE} or \code{FALSE}. Whether or not the lower bounds of vertical axis are determined automatically.}
}
\description{
Draw an FROC  curves and an AFROC curves for user's specified modality and user's specified reader.
Using this function \strong{repeatedly}, we can draw the different reader and modality in a  \strong{same} plane simultaneously.
So, we can visualize the difference of modality (reader).

   --------   To read the tables in Stan S4 class  ----------------------------


  * The  AUC denoted by AA[modalityID , readerID] are shown.

  * The column of 2.5\% and 97.5\% means the lower and upper bounds of the 95% Credible Interval of AUCs.

  * For example, AA[2,3] means the AUC of the 2 nd modality and the 3 rd reader.
}
