% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_error_srsc.R
\name{error_srsc_variance_visualization}
\alias{error_srsc_variance_visualization}
\title{Visualization Of variance Analysis}
\usage{
error_srsc_variance_visualization(
  return.value.of_error_srsc,
  log_scale_x.axis = TRUE
)
}
\arguments{
\item{return.value.of_error_srsc}{A return value
of the function \code{\link{error_srsc}()}.}

\item{log_scale_x.axis}{A logical, whether x axis is log scale.}
}
\value{
A long format dataframe of error and its parameter name
}
\description{
Visualization Of variance Analysis
}
\examples{
\dontrun{
 a <- error_srsc()

 error_srsc_variance_visualization(a)



  a <- error_srsc(replicate.datset = 10)
  error_srsc_variance_visualization(a)

}

}
