% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explanation_about_package_BayesianFROC.R
\name{explanation_about_package_BayesianFROC}
\alias{explanation_about_package_BayesianFROC}
\title{Explanation of this package}
\usage{
explanation_about_package_BayesianFROC()
}
\description{
In \R console, explanation are shown.
}
\examples{

explanation_about_package_BayesianFROC()

}
