% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explanation_for_what_curves_are_drawn.R
\name{explanation_for_what_curves_are_drawn}
\alias{explanation_for_what_curves_are_drawn}
\title{Print out about what curves are drawn}
\usage{
explanation_for_what_curves_are_drawn(modalityID, readerID)
}
\arguments{
\item{modalityID}{A vector.}

\item{readerID}{A  vector..}
}
\value{
Nothing
}
\description{
For package developer.
}
\examples{
 \dontrun{

#================The first example======================================


  modalityID <-c(1,2)
  readerID <-c(1,2,3)

  explanation_for_what_curves_are_drawn( modalityID, readerID )


#================The second example======================================


  modalityID <- 1
  readerID <-c(1,2,3)

  explanation_for_what_curves_are_drawn( modalityID, readerID )


}# dottest


}
