% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data_frame_from_dataList_MRMC.R
\name{extract_data_frame_from_dataList_MRMC}
\alias{extract_data_frame_from_dataList_MRMC}
\title{Extract sub data frame from list of FROC data}
\usage{
extract_data_frame_from_dataList_MRMC(dataList)
}
\arguments{
\item{dataList}{MRMC data list.}
}
\value{
A data frame consisting of  modality ID vector, reader ID vector, confidence vector,
hit vector and false alarms vector.
}
\description{
Extract sub data frame from list of FROC data
}
\examples{
\donttest{


#----------------------------------------------------------------------------------------
#            2)          use exsisting dataset, named dddddd
#----------------------------------------------------------------------------------------



 fit_GUI_Shiny_MRMC(DF=extract_data_frame_from_dataList_MRMC(dddddd))





}

}
