% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replicate_MRMC_data.R
\name{false_and_its_rate_creator}
\alias{false_and_its_rate_creator}
\title{MRMC False Alarm Creator.}
\usage{
false_and_its_rate_creator(z.truth = BayesianFROC::z_truth, NI = 333,
  NL = 111, ModifiedPoisson = FALSE, seed = 12345)
}
\arguments{
\item{z.truth}{Vector of dimension = C represents the thresholds of bi-normal assumption.}

\item{NI}{The number of images.}

\item{NL}{The number of lesions.}

\item{ModifiedPoisson}{Logical, that is \code{TRUE} or \code{FALSE}. If
\code{ModifiedPoisson = TRUE}, then Poisson rate of false alarm are \code{per lesion},
and if \code{ModifiedPoisson = FALSE}, then Poisson rate of false alarm are \code{per image}.
To know detail, refer the author's paper in which I explained per image and per lesion.}

\item{seed}{The seed for creating false alarm which are generated by the Poisson  distributions using the specified seed.}
}
\description{
From threshold, mean and S.D., data of False Alarm are created.
}
\details{
From threshold, mean and S.D.
of the latent Gaussian noise distribution in the bi-normal assumption,
data of False Alarm are created.
For the process of this drawing false alarm samples, its rate are also created. So, in the return values of the function, the rates for each confidence level is also attached.
}
\examples{
 \donttest{
false.rate <- false_and_its_rate_creator()


}
}
