% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fffaaabbb.R
\name{fffaaabbb}
\alias{fffaaabbb}
\title{Package Development tools and memo.}
\usage{
fffaaabbb(a = 0, b = 4, c = 1)
}
\arguments{
\item{a, b, c}{indicating version
project option build and reload
^
^
^
^}
}
\description{
This is for the author of this package.
}
