% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_GUI.R
\name{fit_GUI}
\alias{fit_GUI}
\title{Fit with GUI via Shiny}
\usage{
fit_GUI(DF = data.frame(h = c(97, 32, 31), f = c(1, 14, 74)),
  outdir = system.file("myapp", package = "BayesianFROC"),
  outfilename = "table")
}
\arguments{
\item{DF}{initial data to be fited}

\item{outdir}{I use \code{  system.file("myapp", package = "BayesianFROC")    }}

\item{outfilename}{I do not know :'-D}
}
\value{
None
}
\description{
First, please execute, then user will understand what it is.
This function is the one of the most important function in this package.
I do not assume the user is familiar with R script but FROC analysis. So, I made this function to provide the Graphical User Interface (GUI) for users.
I hope it helps someone in the world.
}
\examples{

\donttest{

 #No need to consider the variables, it is sufficient in  default values.
 fit_GUI()

}

}
