% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_GUI_dashboard.R
\name{fit_GUI_dashboard}
\alias{fit_GUI_dashboard}
\title{Fit with GUI via Shiny (Simple version)}
\usage{
fit_GUI_dashboard(
  DF = data.frame(h = c(97L, 32L, 31L), f = c(1L, 14L, 74L)),
  NL.max = 1111,
  NI.max = 1111,
  NL.initial = 259,
  MCMC.chains.max = 4
)
}
\arguments{
\item{DF}{A dataframe as an initial data to be fitted a model}

\item{NL.max}{max number of bins indicating the maximal number in which the number of lesions can move}

\item{NI.max}{max number of bins indicating the maximal number in which the number of imagegs can move}

\item{NL.initial}{Natural number indicating the initial number of lesions, Default value =259.}

\item{MCMC.chains.max}{max number of bins indicating number of MCMC chains}
}
\value{
None
}
\description{
simple is vest
}
\details{
First, please execute, then user will understand what it is.
This function is the one of the most important function in this package.
I do not assume that the user is familiar with R script but FROC analysis. So, I made this function to provide the Graphical User Interface (GUI) for users.
I hope it helps someone in the world.
}
\examples{

\dontrun{
## Only run examples in interactive R sessions
if (interactive()) {

#========================================================================================
#            1)           Use the default User Interface
#========================================================================================
#'

 #No need to consider the variables, it is sufficient in  default values.


 #fit_GUI_dashboard()




#========================================================================================
#            2)           Change the  User Interface
#========================================================================================


#  We can change the max imput of the number of lesions and the max of number of images
#

 #fit_GUI_dashboard(NL.max = 2222,
 #               NI.max = 3333)





#========================================================================================
#            3)           Change the  Default value
#========================================================================================



#  fit_GUI_dashboard(
#  DF= data.frame( h=dataList.Chakra.4$h,
#                  f=dataList.Chakra.4$f
#                )
#                )

# Or equivalently,

#  fit_GUI_dashboard(
#             DF= data.frame(
#             h = c(160,  25,  15,   7),
#             f = c(  8,  16,  18,  13)
#                          )
#             )


#========================================================================================
#            4)           Change the user Imterface
#========================================================================================



#fit_GUI_dashboard(

#           DF= data.frame(
#               h = c(160,  25,  15,   7),
#               f = c(  8,  16,  18,  13)
#                 ),

#                 NL.max = 1192,
#                 NI.max = 794,
#                 MCMC.chains.max = 6

#               )





}### Only run examples in interactive R sessions


}

}
\author{
Issei Tsunoda
}
