% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_GUI.R
\name{fit_GUI_MRMC}
\alias{fit_GUI_MRMC}
\title{Fit with GUI via Shiny}
\usage{
fit_GUI_MRMC()
}
\value{
None
}
\description{
First, please execute, then user will understand what it is.
This function is the one of the most important function in this package.
I do not assume the user is familiar with R script but FROC analysis. So, I made this function to provide the Graphical User Interface (GUI) for users.
I hope it helps someone in the world.
}
