% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_GUI_MRMC_new.R
\name{fit_GUI_MRMC_new}
\alias{fit_GUI_MRMC_new}
\title{Fit an MRMC model to data with Shiny GUI}
\usage{
fit_GUI_MRMC_new(M = 2, Q = 3, C = 4)
}
\arguments{
\item{M}{mo}

\item{Q}{re}

\item{C}{con}
}
\value{
ret
}
\description{
I love you.
}
\details{
I need you.
}
