% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_GUI_simple_from_apppp_file.R
\name{fit_GUI_simple_from_apppp_file}
\alias{fit_GUI_simple_from_apppp_file}
\title{Fit with GUI via Shiny}
\usage{
fit_GUI_simple_from_apppp_file(display.mode = FALSE)
}
\arguments{
\item{display.mode}{Logical, passing to \code{runApp}.
Default is \code{FALSE} corresponding to "normal",
 and if \code{TRUE}, then "showcase" which shows code.
 The author made this, but it did not work or ignored, that is, showcase did not work.
 Why???}
}
\value{
None
}
\description{
First, please execute, then user will understand what it is.
This function is the one of the most important function in this package.
I do not assume  that the user is familiar with R script but FROC analysis. So, I made this function to provide the Graphical User Interface (GUI) for users.
I hope it helps someone in the world.
}
\examples{

\donttest{

 #No need to consider the variables, it is sufficient in  default values.
 #fit_GUI_simple_from_apppp_file()

}#'


}
\author{
Issei Tsunoda
}
