% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc_new.R
\name{flatnames}
\alias{flatnames}
\title{from rstan package}
\usage{
flatnames(names, dims, col_major = FALSE)
}
\arguments{
\item{names}{A vector of characters}

\item{dims}{A positive integer}

\item{col_major}{A logical}
}
\value{
A vector of characters
}
\description{
from rstan package
}
\examples{

#  flatnames(c("a","b"),3)

# [1] "a[1]" "a[2]" "a[3]" "b[1]" "b[2]" "b[3]"


}
\author{
Some Stan developer, I am not sure,..., who?
}
