% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FROC_via_ggplot.R
\name{FROC_curve}
\alias{FROC_curve}
\title{FROC curve as an embedding map}
\usage{
FROC_curve(x)
}
\arguments{
\item{x}{A real number moves in domain  of FROC curve}
}
\value{
none
}
\description{
FROC curve as an embedding map
}
\examples{

# I love you!



}
