% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posterior_variance.R
\name{get_posterior_variance}
\alias{get_posterior_variance}
\title{(UNDER CONSTRUCTION) Alternative of \code{rstan::get_posterior_mean()}}
\usage{
get_posterior_variance(StanS4class, name.of.parameter)
}
\arguments{
\item{StanS4class}{An S4 object of class \emph{\code{ \link{stanfitExtended}}} which is an inherited class from the S4 class  \code{\link[rstan]{stanfit}}.
This \R object is a fitted model object
 as a return value of the function \code{\link{fit_Bayesian_FROC}()}.

It can be passed to \code{\link{DrawCurves}()}, \code{\link{ppp}()}  and ... etc}

\item{name.of.parameter}{An parameter name (given as a character string, should not surround by "").
The name of parameter which user want to extract.
Parameters are contained in the parameter block of each Stan file in the path: inst/extdata.}
}
\value{
variance or posterior parameters, if it is an array, then return is also an array.
}
\description{
This function is underconstruction. I validate only the example of this function.
For MRMC case, I have to write or modify code. 2019 Sept 6
}
\examples{



\donttest{

        fit <- fit_Bayesian_FROC(BayesianFROC::dd)




             e <- rstan::extract(fit)




 # Check the retrun value is desired one.


   apply(e$z, 2, var) ==  get_posterior_variance(fit,z)

             var(e$m) ==  get_posterior_variance(fit,m)

             var(e$v) ==  get_posterior_variance(fit,v)






 }#donttest





}
