% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_treedepth_threshold.R
\name{get_treedepth_threshold}
\alias{get_treedepth_threshold}
\title{get treedepth threshold}
\usage{
get_treedepth_threshold(StanS4class)
}
\arguments{
\item{StanS4class}{An S4 object of class \emph{\code{ \link{stanfitExtended}}} which is an inherited class from the S4 class  \code{stanfit}.
This \R object is a fitted model object
 as a return value of the function \code{\link{fit_Bayesian_FROC}()}.

To be passed to \code{\link{DrawCurves}()}, \code{\link{ppp}()}  and ... etc}
}
\value{
A non-negative integer
}
\description{
From \code{rstan:::get_treedepth_threshold}.
}
\author{
Some Stan developer. Not the author of this package!
}
