% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hit_generator_from_multinomial.R
\name{hit_generator_from_multinomial}
\alias{hit_generator_from_multinomial}
\title{Under Const}
\usage{
hit_generator_from_multinomial(
  z.truth = c(0.1, 0.2, 0.3, 0.4, 0.5),
  mu.truth = 1,
  v.truth = 2
)
}
\arguments{
\item{z.truth}{Vector of dimension = C represents the thresholds of bi-normal assumption.}

\item{mu.truth}{array of dimension (M,Q). Mean of the signal distribution of bi-normal assumption.}

\item{v.truth}{array of dimension (M,Q). Standard Deviation of represents the signal distribution of bi-normal assumption.}
}
\value{
A vector of non-negative integers
}
\description{
Under Const
}
\details{
The algorithm of \code{rmultinom()} explained in \code{?rmultinom} is quite same as mine code.
So, I do not need to write this code. OK.
}
