% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hit_rate_adjusted_from_the_vector_p.R
\name{hit_rate_adjusted_from_the_vector_p}
\alias{hit_rate_adjusted_from_the_vector_p}
\title{hit rate adjusted from a vector p}
\usage{
hit_rate_adjusted_from_the_vector_p(p_vector)
}
\arguments{
\item{p_vector}{A vector}
}
\value{
A vector
}
\description{
hit rate adjusted from a vector p
}
\examples{

 p <- c(1,2,3)
 a <- hit_rate_adjusted_from_the_vector_p( p )
 a

 # [1] -0.25 -1.00  3.00

 a[3] == 3
 a[2] == p[2]/(1-p[3])
 a[1] == p[1]/(1-p[3]-p[2])



#========================================================================================
#                               application in the function ppp_srsc in this package
#========================================================================================

\dontrun{


 f <- fit_Bayesian_FROC( dataList = d )
 e <-rstan::extract(f)

q<-e$p[1,]
hit_rate_adjusted_from_the_vector_p(q)
t(apply(e$p,hit_rate_adjusted_from_the_vector_p,MARGIN = 1))[1,]
q<-e$p[2,]
hit_rate_adjusted_from_the_vector_p(q)
t(apply(e$p,hit_rate_adjusted_from_the_vector_p,MARGIN = 1))[2,]



}



}
