% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/validation_MRMC_Create_dataList_MRMC_Hit_from_rate_etc.R
\name{hits_from_thresholds}
\alias{hits_from_thresholds}
\title{MRMC Hit  Creator from thresholds, mean and S.D.}
\usage{
hits_from_thresholds(
  z.truth = BayesianFROC::z_truth,
  mu.truth = BayesianFROC::mu_truth,
  v.truth = BayesianFROC::v_truth,
  NL = 252,
  seed = 123
)
}
\arguments{
\item{z.truth}{Vector of dimension = C represents
the thresholds of bi-normal assumption.}

\item{mu.truth}{array of dimension (M,Q).
Mean of the signal distribution
of bi-normal assumption.}

\item{v.truth}{array of dimension (M,Q).
Standard Deviation of represents the signal
 distribution of bi-normal assumption.}

\item{NL}{Number of Lesions.}

\item{seed}{The seed for creating data consisting of the number of hits synthesized by the binomial distributions with the specified seed.}
}
\value{
Hits Data for MRMC. The reason that hits
is multiple reader and multiple modalities arise
from the multiple indices of
mean and S.D. of signal distribution of the
bi-normal assumption.
}
\description{
From threshold, mean and S.D.,
 data of hit rate are created.
}
\examples{
\donttest{
hits.rate.p <-hits_from_thresholds()


}#donttest
}
