% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replicate_MRMC_data.R
\name{hits_rate_creator}
\alias{hits_rate_creator}
\title{MRMC Hit Rates Creator from Thresholds, Mean and S.D.}
\usage{
hits_rate_creator(z.truth = BayesianFROC::z_truth,
  mu.truth = BayesianFROC::mu_truth, v.truth = BayesianFROC::v_truth)
}
\arguments{
\item{z.truth}{Vector of dimension = C represents the thresholds of bi-normal assumption.}

\item{mu.truth}{array of dimension (M,Q). Mean of represents the signal distribution of bi-normal assumption.}

\item{v.truth}{array of dimension (M,Q). Standard Deviation of represents the signal distribution of bi-normal assumption.}
}
\value{
ppp:= hit rate, which is an array with dimension
}
\description{
From thresholds, data of hit rate are created.
}
\examples{

\donttest{
#================The first example======================================

#     Using default values for hit rates, we can create a data of hits as follows:

        hits.rate <-hits_rate_creator()

#================The second example======================================

#     Using the hit rate from the hits_rate_creator(), we can get the hits data:

        hits_creator_from_rate(p.truth =hits_rate_creator() )

#================The remark for example======================================

# The author does not show how to specify the hit rates or threshods.
# For the details of it, please see the default values of such a quantities.


#================The 4-th example======================================

    p.truth.array <- hits_rate_creator()

}# dottest

}
