% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertical_to_horizontal.R
\name{horizontal_from_vertical_in_each_case}
\alias{horizontal_from_vertical_in_each_case}
\title{Transfer From Vertical placement into Horizontal placement for casewise vectors}
\usage{
horizontal_from_vertical_in_each_case(vector, NI, M, Q, C)
}
\arguments{
\item{vector}{a vector, such as hit vector, h or f counted by each cases.}

\item{NI}{Number of images, in other words, cases}

\item{M}{Number of modalities}

\item{Q}{Number of readers}

\item{C}{Number of confidence levels, in other words, ratings}
}
\value{
an array whose dimension is NI, M,Q,C
}
\description{
Transfer From Vertical placement into Horizontal placement for casewise vectors
}
\examples{

h <-dcasewise$h
horizontal_from_vertical_in_each_case(h, 200,5,4,5)

f <-dcasewise$f
horizontal_from_vertical_in_each_case(f, 200,5,4,5)
}
