% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_imports.R
\name{install_imports}
\alias{install_imports}
\title{Installer.}
\usage{
install_imports()
}
\description{
This is an installer for  required packages in this package.
To install this package BayesianFROC, we use the package xlsx which require the Java. So, if use buy a new computer and and it does not have installed the Java, then please install Java.
}
