% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_logical_0.R
\name{is_length_zero}
\alias{is_length_zero}
\title{Is argument of length zero ?}
\usage{
is_length_zero(integer_object)
}
\arguments{
\item{integer_object}{An object of class integer}
}
\value{
A logical
}
\description{
When object is created by the codes \code{ x <- integer(); y <- list(); z <- logical()},
and if the values is not substituted, then this function return \code{TRUE}.
This function determine whether the value is assigned or not according to the object size.

 2020 Oct 6
}
\examples{

a <- integer()

is_length_zero(a)

is_length_zero(1)

a <- list()

is_length_zero(a)

is_length_zero(TRUE)


}
