% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_logical_0.R
\name{is_logical_0}
\alias{is_logical_0}
\title{is.logical(0)}
\usage{
is_logical_0(integer_object)
}
\arguments{
\item{integer_object}{An object of class integer}
}
\value{
A logical
}
\description{
When object is created by the codes \code{ x <- integer(); y <- list(); z <- logical()},
and if the values is not substituted, then this function return \code{TRUE}.
This function determine whether the value is assigned or not according to the object size.

 2020 Sept 25
}
\examples{

a <- integer()

is_logical_0(a)

is_logical_0(1)

a <- integer()

is_logical_0(a)

is_logical_0(TRUE)


}
