% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_plot.R
\name{is_stanfitExtended}
\alias{is_stanfitExtended}
\title{Check whether class is \emph{stanfitExtended} for any \R object}
\usage{
is_stanfitExtended(any_R_object)
}
\arguments{
\item{any_R_object}{any R object}
}
\value{
logical
}
\description{
Check whether class is \emph{stanfitExtended} for any \R object
}
