% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_q_c_vector_from_M_Q_C.R
\name{m_q_c_vector_from_M_Q_C}
\alias{m_q_c_vector_from_M_Q_C}
\title{Creats vectors: \code{m,q,c}  from integers: \code{M,Q,C}}
\usage{
m_q_c_vector_from_M_Q_C(M, Q, C)
}
\arguments{
\item{M}{A positive integer, representing modality ID}

\item{Q}{A positive integer, representing reader ID}

\item{C}{A positive integer, representing confidence level}
}
\value{
A data-frame, including three vectors, which are named \code{m,q,c}  representing modality ID and reader ID and confidence level, respectively.

For example, the resulting object of \code{a <- m_q_c_vector_from_M_Q_C(2,3,4)} is given by

\code{  > a}


\tabular{rrr}{
  \code{ m }\tab \code{ q} \tab  \code{c} \cr
  -------- \tab ------ \tab ---------\cr
  1 \tab 1 \tab 4\cr
  1 \tab 1 \tab 3\cr
  1 \tab 1 \tab 2\cr
  1 \tab 1 \tab 1\cr
  1 \tab 2 \tab 4\cr
  1 \tab 2 \tab 3\cr
  1 \tab 2 \tab 2\cr
  1 \tab 2 \tab 1\cr
  1 \tab 3 \tab 4\cr
  1 \tab 3 \tab 3\cr
  1 \tab 3 \tab 2\cr
  1 \tab 3 \tab 1\cr
  2 \tab 1 \tab 4\cr
  2 \tab 1 \tab 3\cr
  2 \tab 1 \tab 2\cr
  2 \tab 1 \tab 1\cr
  2 \tab 2 \tab 4\cr
  2 \tab 2 \tab 3\cr
  2 \tab 2 \tab 2\cr
  2 \tab 2 \tab 1\cr
  2 \tab 3 \tab 4\cr
  2 \tab 3 \tab 3\cr
  2 \tab 3 \tab 2\cr
  2 \tab 3 \tab 1
}
}
\description{
Makes \code{m,q,c} vectors
 from a collection of three integers \code{M,Q,C},
 where three vectors \code{m,q,c}
 denotes modality ID, reader ID, confidence level,
 respectively.
}
\details{
My research is not supported any found, I am completely independent
 and only my own or my parents are supported my research.
 No internet, poor condition, I made this. I must go on untill jounal accepts my manuscripts.

 I am not happy to spent with FROC analysis, since it is not my interest.
 I want to research pure mathematics. I do not want to waste a time.
 I do not want to waste a time in hospital or plurigo nodularis.
 When I become happy? This program helps me? With great pain at 2019 Sept.
 2019 Sept. 8
}
\examples{


#========================================================================================
#                     Create a ID vectors
#========================================================================================

  a <- m_q_c_vector_from_M_Q_C(2,3,4)

  a$m
  a$q
  a$c



#========================================================================================
#                      validation of this function
#========================================================================================
#'


            a   <- m_q_c_vector_from_M_Q_C(5,4,5)


            a$m == dd$m
            a$c == dd$c
            a$q == dd$q




}
