% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_TeX_file_for_summary.R
\name{make_TeX}
\alias{make_TeX}
\title{Make a TeX file for summary}
\usage{
make_TeX()
}
\value{
TeX file reflected the analysis
}
\description{
Under Construction...
``This only inner funtion, in the future I run this in the \code{\link{fit_Bayesian_FROC}()}.
}
