% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_true_parameter_MRMC.R
\name{make_true_parameter_MRMC}
\alias{make_true_parameter_MRMC}
\title{Make a true model parameter and include it in this package}
\usage{
make_true_parameter_MRMC(StanS4class)
}
\arguments{
\item{StanS4class}{An S4 object of class \emph{\code{ \link{stanfitExtended}}} which is an inherited class from the S4 class  \code{\link[rstan]{stanfit}}  that can be passed to the \code{\link{DrawCurves}()}, \code{\link{ppp}()}  and ... etc}
}
\description{
Make a true model parameter and include it in this package
}
