% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata_to_DrawCurve_MRMC}
\alias{metadata_to_DrawCurve_MRMC}
\title{Create metadata for MRMC data}
\usage{
metadata_to_DrawCurve_MRMC(StanS4class, mesh.for.drawing.curve = 5000)
}
\arguments{
\item{StanS4class}{An S4 object of class \emph{\code{ \link{stanfitExtended}}} which is an inherited class from the S4 class  \code{stanfit}.
This \R object is a fitted model object
 as a return value of the function \code{\link{fit_Bayesian_FROC}()}.

To be passed to \code{\link{DrawCurves}()} ... etc}

\item{mesh.for.drawing.curve}{A positive large integer, indicating number of dots drawing the curves, Default =10000.}
}
\value{
A metadata such as number of cumulative false alarms and hits to create and draw the curve.
}
\description{
From data of number of hits and false alarms, we calculate the number of cumulative false positives and hits.
Since there are three subscripts, reader, modality, and image, we create array format and vector format etc...
}
