% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_true_param.R
\docType{data}
\name{mu}
\alias{mu}
\title{Mean of signal: parameter of an MRMC model}
\description{
A posterior mean of the model parameter for data \code{\link{ddd}} as a truth parameter example.
}
\examples{


#> BayesianFROC::mu

#[,1]      [,2]     [,3]      [,4]
#[1,] 1.914686 0.7933306 1.526482 0.9543375
#[2,] 2.008008 1.2005846 2.081756 1.0197703
#[3,] 1.532117 0.5851726 1.513018 0.8879678


# [modality, reader]





}
\seealso{
\code{make_true_parameter_MRMC}
}
\author{
Issei Tsunoda \email{tsunoda.issei1111@gmail.com }
}
