% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_true_param.R
\docType{data}
\name{mu_truth}
\alias{mu_truth}
\title{Mean of signal: parameter of an MRMC model}
\description{
A posterior mean of the model parameter for data \code{\link{ddd}} as a truth parameter example.
}
\details{
Mean Rate data of some MRMC data to use as a default value
of the function \code{ hits_creator_from_rate}.
This is an array obtained from estimates of some data contained in this package.
To simulate a replication of dataset, the default values should be used from an actual values.
 Thus the author prepare this data.
}
\examples{


#> mu_truth
#
#         [,1]      [,2]     [,3]      [,4]
#[1,] 1.730751 0.8298189 1.334771 0.6386057
#[2,] 1.812523 1.1889223 1.883562 0.7185546
#[3,] 1.319588 0.6062924 1.248589 0.5458920


# [modality, reader]





}
\seealso{
\code{hits_creator_from_rate}
}
\author{
Issei Tsunoda \email{tsunoda.issei1111@gmail.com }
}
