% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_value_visualization.R
\name{p_value_visualization}
\alias{p_value_visualization}
\title{Calculation of P values are visualized}
\usage{
p_value_visualization(return_of_ppp_srsc)
}
\arguments{
\item{return_of_ppp_srsc}{An object of return value of \code{ppp_srsc}()}
}
\value{
NaN
}
\description{
The author suspect the p value calculation in this Pkg is wrong, cuz it is always accept the null hypothesis.
Why? So,,, I made this to examine the process of comparison between observed chisquare and that of post. pred. dist.
Hahh, why? My intuition is wrong? or  program is incorrect?? I'm not gonna write program.
}
\examples{

 \dontrun{
  f <- fit_Bayesian_FROC(
       ite  = 55,
        summary = TRUE,
        cha = 1,
         dataList = dataList.Chakra.1
         )

  a<- ppp_srsc(f)

  p_value_visualization(a)

}


}
