% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs_plot_if_divergent_transition_occurred.R
\name{pairs_plot_if_divergent_transition_occurred}
\alias{pairs_plot_if_divergent_transition_occurred}
\title{Pairs plot for divergent transition}
\usage{
pairs_plot_if_divergent_transition_occurred(StanS4class,
  character.representing.paramter = "z")
}
\arguments{
\item{StanS4class}{An \R  object of class \emph{\code{ \link{stanfitExtended}}} inherited from the S4 class  \code{\link[rstan]{stanfit}}  that can be passed to the \code{\link{DrawCurves}()} and  \code{\link{p_value_of_the_Bayesian_sense_for_chi_square_goodness_of_fit}()}, ..., etc.}

\item{character.representing.paramter}{Character, surrounded by "", indicating the paramter of model.}
}
\description{
If divergent transition occurs, the author often forget the variable par or pars. So, I made this to avoid such confusion.
}
\examples{
\donttest{


# Create a fitted model object of class stanfitExtended  inherited from stanfit.

 fit <- fit_Bayesian_FROC( ite  = 1111,
                           summary = FALSE,
                           cha = 1,
                           Null.Hypothesis = FALSE,
                           dataList = dd )

# Pairs plot to examine the divergent transition.


   pairs_plot_if_divergent_transition_occurred(fit)

    }


}
