% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phi__and__Phi_inv.R
\name{Phi_inv}
\alias{Phi_inv}
\title{Inverse function of the Cumulative distribution
 function \eqn{\Phi(x)} of the Standard Gaussian.
 where \eqn{x} is a  real number.}
\usage{
Phi_inv(x)
}
\arguments{
\item{x}{A real. To be passed
to the function \code{stats::qnorm()}}
}
\value{
A real number: \eqn{\Phi^{-1}(x)}
}
\description{
The author is confused  \code{stats::qnorm()}
 with  \code{stats::pnorm()} and thus he made this.
}
\details{
In Stan file, it is \code{inv_Phi()}
and not \code{inv_phi}.

Since  \eqn{\Phi(x)} is monotonic, it follows that
 \eqn{\frac{d}{dx}\Phi^{-1} = (\frac{d}{dx}\Phi)^{-1}  >0},
 and thus  \eqn{\Phi^{-1}(x)} is also monotonic.
}
\examples{




         x <- runif(100)

 Phi_inv(x) == stats::qnorm(x)

 inv_Phi(x) == stats::qnorm(x)






}
\seealso{
\code{\link{Phi}()},  \code{\link{inv_Phi}()}
}
