% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{plot,stanfitExtended,missing-method}
\alias{plot,stanfitExtended,missing-method}
\title{This is a method for a generic function print}
\usage{
\S4method{plot}{stanfitExtended,missing}(x)
}
\arguments{
\item{x}{This is an object of an S4 class named stanfitExtended which is an inherited S4 class from the stanfit S4 class in the rstan package.}
}
\description{
This is a method for print and stanfitExtended S4 class.
}
