% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp_vectorization.R
\name{plot_dataset_of_ppp_MRMC}
\alias{plot_dataset_of_ppp_MRMC}
\title{plot datasets using calculation of ppp}
\usage{
plot_dataset_of_ppp_MRMC(StanS4class, summary = FALSE)
}
\arguments{
\item{StanS4class}{An S4 object of class \emph{\code{ \link{stanfitExtended}}} which is an inherited class from the S4 class  \code{stanfit}.
This \R object is a fitted model object
 as a return value of the function \code{\link{fit_Bayesian_FROC}()}.

To be passed to \code{\link{DrawCurves}()} ... etc}

\item{summary}{Logical: \code{TRUE} of \code{FALSE}. Whether to print the verbose summary. If \code{TRUE} then verbose summary is printed in the \R console. If \code{FALSE}, the output is minimal. I regret, this variable name should be verbose.}
}
\value{
p value whose null hypothesis is that model is fitted to data well.
}
\description{
plot datasets using calculation of ppp
}
\examples{
\dontrun{
#========================================================================================
#                              Now single reader and single modality case only
#========================================================================================
# Fit a model to data-set "dd"
# In the resulting object contained samples from posterior predictive distribution (PPD)

f <- fit_Bayesian_FROC( ite  = 1111, summary = TRUE,  cha = 1, dataList = dd )


# Plot samples synthesized from posterior predictive distribution (PPD)

plot_dataset_of_ppp_MRMC(f)

}#dontrun

}
