% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFROC.R
\name{plotFROC}
\alias{plotFROC}
\title{Draw FROC curves by two parameters a and b}
\usage{
plotFROC(a, b, mesh.for.drawing.curve = 10000, upper_x = 1,
  upper_y = 1, lower_y = 0)
}
\arguments{
\item{a}{See paper for definition.}

\item{b}{See paper for definition.}

\item{mesh.for.drawing.curve}{This is a number of points using draw the curves, default =10000}

\item{upper_x}{The frame size of drawing picture.}

\item{upper_y}{The frame size of drawing picture.}

\item{lower_y}{The frame size of drawing picture.}
}
\description{
Plot FROC curves based on two parameters a and b.
}
