% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnorm_or_qnorm.R
\name{pnorm_or_qnorm}
\alias{pnorm_or_qnorm}
\title{pnorm or qnorm}
\usage{
pnorm_or_qnorm()
}
\description{
Thu author always forget which is cumulative distribution of Gaussia, so I made this and this tells me which is mmy desired one.
In this package, I often use \eqn{\Phi()} for the standard Gaussian, and it is \code{pnorm()}.
I am very confuse, since probability density  has initial alphabet p, but \code{pnorm()} is not it.
}
