% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp.R
\name{ppp_MRMC}
\alias{ppp_MRMC}
\title{MRMC: Posterior Predictive P value (PPP) for MRMC,}
\usage{
ppp_MRMC(
  StanS4class,
  summary = TRUE,
  replicate.number.from.model.for.each.MCMC.sample = 2
)
}
\arguments{
\item{StanS4class}{An S4 object of class \emph{\code{ \link{stanfitExtended}}} which is an inherited class from the S4 class  \code{stanfit}.
This \R object is a fitted model object
 as a return value of the function \code{\link{fit_Bayesian_FROC}()}.

To be passed to \code{\link{DrawCurves}()}, \code{\link{ppp}()}  and ... etc}

\item{summary}{Logical: \code{TRUE} of \code{FALSE}. Whether to print the verbose summary. If \code{TRUE} then verbose summary is printed in the \R console. If \code{FALSE}, the output is minimal. I regret, this variable name should be verbose.}

\item{replicate.number.from.model.for.each.MCMC.sample}{A positive integer, representing \eqn{J} in the following notation.}
}
\value{
A positive number indicates Posterior Predictive P value (ppp).
}
\description{
PPP for chi square goodness of fit statistic
}
\details{
The author hates the notion of p value and this is the motivation that he developed new theory without p values.
However, he cannot overcome the traditional people. he loves mathematics, but he hates statistics.
he emphasizes that notion of p value is dangerous (monotonicity w.r.t. sample size) and its background is unknown.
Of course, intuitively, it is good. But, the theoritically, it does not ensure some criterion in large sample comtext.

So, p value said that my effort is rarely admissible, since its p value said that he is small for various datasets.
So, this funcking p value said my effort is wrong, or should change  model.
Unfortunately, my hand aches cannot program more models.
 Ha,... why many peoply like p value bitch.
}
\examples{



\dontrun{
#========================================================================================
#  1)  Fit a Model to MRMC Data
#========================================================================================

  fit <- fit_Bayesian_FROC( ite  = 33,  dataList = ddd )

#========================================================================================
#  1)  Evaluate Posterior Predictive P value for the Goodness of Fit
#========================================================================================

     ppp_MRMC(fit)

#  If this quantity is greater, then we may say that our model is better.

#  I made this ppp at 2019 August 25.

     Close_all_graphic_devices() # 2020 August
}#'




}
