% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp.R
\name{ppp_srsc}
\alias{ppp_srsc}
\title{Calculates PPP for Models of a single reader and a single modality (Calculation is correct! :'-D)}
\usage{
ppp_srsc(StanS4class, Colour = TRUE, dark_theme = TRUE, plot = TRUE,
  summary = TRUE,
  replicate.number.from.model.for.each.MCMC.sample = 100)
}
\arguments{
\item{StanS4class}{An S4 object of class \emph{\code{ \link{stanfitExtended}}} which is an inherited class from the S4 class  \code{\link[rstan]{stanfit}}  that can be passed to the \code{\link{DrawCurves}()}, \code{\link{ppp}()}  and ... etc}

\item{Colour}{Logical: \code{TRUE} of \code{FALSE}.  whether Colour of curves is dark theme or not.}

\item{dark_theme}{TRUE or FALSE}

\item{plot}{Logical, whether replicated data are drawn, in the following notation, replicated data are denoted by \eqn{y_1,y_2,...,y_N}.}

\item{summary}{Logical: \code{TRUE} of \code{FALSE}. Whether to print the verbose summary, i.e., logical; If \code{TRUE} then verbose summary is printed in the \R console. If \code{FALSE}, the output is minimal. I regret, this variable name should be verbose.}

\item{replicate.number.from.model.for.each.MCMC.sample}{A positive integer, representing \eqn{J} in the following notation.
Now, I think all I needed is love! ttu ttu tututu Love is all I need.

Suppose that  \deqn{\theta_1, \theta_2, \theta_3,...,\theta_n}   is drawn from posterior \eqn{\pi(\theta|D)} of given data \eqn{D}.

Let \eqn{y_1,y_2,...,y_n} be samples drawn from

\deqn{y_1 ~ likelihood ( . |\theta_1), }
\deqn{y_2 ~ likelihood ( . |\theta_2),}
\deqn{y_3 ~ likelihood ( .|\theta_3),}
\deqn{...,}
\deqn{y_n ~ likelihood ( . |\theta_N),}



Then the list of return values retains the following:
\describe{
\item{ \code{chisq_at_observed_data}      }{ \deqn{\chi (D|\theta_1), \chi (D|\theta_2), \chi (D|\theta_3),...,\chi (D|\theta_n),} }
\item{ \code{chisq_not_at_observed_data}  }{\deqn{\chi (y_1|\theta_1), \chi (y_2|\theta_2), \chi (y_3|\theta_3),...,\chi (y_n|\theta_n),  }}
\item{ \code{Logical}                     }{ The i-th component is a logical vector indicating whether  \deqn{\chi (y_2|\theta_2) > \chi (D|\theta_2)} is satisfied or not. Oppai ga Ippai. If \code{TRUE}, then the inequality holds.}
\item{ \code{p.value}                     }{  From the component \code{Logical}, we calculate the so-called \emph{Posterior Predictive P value}. Note that the author hate this notion!! I hate it!! Akkan Beeeee!!! }
}}
}
\value{
A list, including p value and materials to calculate it.
}
\description{
Calculates Posterior Predictive P value for chi square (goodness of fit)
}
\details{
Also, plot a data drawn from likelihood with each MCMC parameters generated by HMC.
That is, let us denote the MCMC samples of size  \eqn{n} by \deqn{\theta_1, \theta_2, \theta_3,...,\theta_n ~ \pi(\theta|D)}  drawn from posterior \eqn{\pi(\theta|D)} of given data \eqn{D}.
That is,
\deqn{\theta_1  ~ \pi(.|D), }
\deqn{\theta_2 ~ \pi(.|D), }
\deqn{\theta_3  ~ \pi(.|D),}
\deqn{...,}
\deqn{\theta_n  ~ \pi(.|D),}

which are drawn using the Hamiltonian Monte Carlo Sampling: HMC.

Then, we plot data \eqn{y_1^1,y_2^1,...,y_I^J} from

\deqn{y_1^1,y_1^2,y_1^3,...,y_1^j,....,y_1^J ~ likelihood ( . |\theta_1),}
\deqn{y_2^1,y_2^2,y_2^3,...,y_2^j,....,y_2^J ~ likelihood ( . |\theta_2),}
\deqn{y_3^1,y_3^2,y_3^3,...,y_3^j,....,y_3^J ~ likelihood ( .|\theta_3), }
\deqn{...,}
\deqn{y_i^1,y_i^2,y_i^3,...,y_i^j,....,y_i^J ~ likelihood ( . |\theta_i),}
\deqn{...,}
\deqn{y_I^1,y_I^2,y_I^3,...,y_I^j,....,y_I^J ~ likelihood ( . |\theta_I),}




Let \eqn{ \chi(y|\theta)  } be a chi square goodness of fit statistics of our hierarchical Bayesian Model


\deqn{\chi(y|\theta) := \sum_{r=1}^R \sum_{m=1}^M \sum_{c=1}^C ( \frac { ( H_{c,m,r}-N_L\times p_{c,m,r})^2}{N_L\times p_{c,m,r}} + \frac{(F_{c,m,r}-(\lambda_{c} -\lambda_{c+1} )\times N_{L})^2}{(\lambda_{c} -\lambda_{c+1} )\times N_{L} }).}


and a chi square goodness of fit statistics of our non-hierarchical Bayesian Model

\deqn{\chi(y|\theta) :=  \sum_{c=1}^C \biggr( \frac{( H_{c}-N_L\times p_{c})^2}{N_L\times p_{c}} + \frac{(F_{c}-(\lambda_{c} -\lambda_{c+1} ) )\times N_{L}]^2}{(\lambda_{c} -\lambda_{c+1} )\times N_{L} }\biggr).}

where  a dataset \eqn{y} denotes \eqn{ (F_{c,m,r}, H_{c,m,r}) } in MRMC case and  \eqn{ (F_{c}, H_{c}) } in a single reader and a single modality case,
and model parameter \eqn{\theta}.


Then we can calculate the \emph{posterior predictive p value} for a given dataset \eqn{y_0}.

\deqn{ \int \int I( \chi(y|\theta) > \chi(y_0|\theta) )   f(y|\theta) \pi(\theta|y_0) d \theta d y  }
\deqn{ \approx \int \sum_i I( \chi(y|\theta_i) > \chi(y_0|\theta_i) )   f(y|\theta_i) d y   }
\deqn{ \approx \sum_{j=1}^J \sum_{i=1}^I I( \chi(y_i^j|\theta_i) > \chi(y_0|\theta_i) )     }



  When we plot these replicated data \eqn{y_i^j}, we use the \code{jitter()} which adds a small amount of noise to \strong{avoid overlapping points}.
For example, \code{jitter(c(1,1,1,1))} returns  values: \code{1.0161940 1.0175678 0.9862400 0.9986126}, which is changed
from \code{1,1,1,1} to be not exactly 1 by adding tiny errors to avoid overlapping. I love you. 2019 August 19
Nowadays, I cannot remove my self from some notion, such as honesty, or pain, or,.. maybe these thing is no longer with myself.
This programm is made to fix previous release calculation. Now, this programm calculates correct p value.

So... I calculate the ppp for MCMC and Graphical User Interface based on Shiny for MRMC, which should be variable such as
number of readers, modalities, to generate such ID vectors automatically. Ha,... tired! Boaring, I want to die...t, diet!!
Tinko, tinko unko unko. Manko manko. ha.

Leberiya, he will be die, ha... he cannot overcome, very old, old guy.
 I will get back to meet him. Or I cannot meet him? Liberiya,...very wisdom guy,
Ary you already die? I will get back with presents for you. Ball, I have to throgh ball, and he will catch it.


The reason why the author made the plot of data drawn from \strong{Posterior Predictive likelihoods with each MCMC parameters} is
to understand our programm is correct, that is, each drawing is very mixed. Ha,.... when wright this,... I always think who read it.
I love you, Ruikobach. Ruikobach is tiny and tiny, but,... cute. Ruikosan...Ruiko...
But he has time only several years. He will die, he lives sufficiently so long, ha.

Using this function, user would get \strong{reliable posterior predictive p values}, Cheers! Pretty Crowd!



We note that the calculation of posterior perdictive p value (PPP) relies on the law of large number.
Thus, in order to obtain the relicable PPP, we need to enough large MCMC samples to approximate
the double integral of PPP.
For example, the MCMC samples is small, then R hat is far from 1 but, the low MCMC samples leads
us to incorrect p value which sometimes said that the model is correct even if the R hat criteria
reject the MCMC results.
}
\examples{


\donttest{

#----------------------------------------------------------------------------------------
#            1) Create a fitted model object with data named only one word d
#----------------------------------------------------------------------------------------



                           fit <- fit_Bayesian_FROC(d )



#----------------------------------------------------------------------------------------
#            2) Calculate p value and meta data
#----------------------------------------------------------------------------------------



                              ppp <- ppp(fit)



#----------------------------------------------------------------------------------------
#            3) Extract a p value
#----------------------------------------------------------------------------------------




                              ppp$p.value


# Revised 2019 August 19 with great love and peace and hate of notion of p values.
}



}
\author{
Issei Tsunoda, Prof. of Curlbus University, Mosquitobus and Gostbus univ. also. My technique of catch mosquitos are execellent, so, I am a prof. ha,, employ me. My health is bad, my life will be over.
}
