% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_minimal_reproducible_code.R
\name{print_minimal_reproducible_code_in_case_of_MRMC}
\alias{print_minimal_reproducible_code_in_case_of_MRMC}
\title{Show minimal code in MRMC}
\usage{
print_minimal_reproducible_code_in_case_of_MRMC()
}
\value{
NULL ?
}
\description{
Now 2020 March, it is available.
}
\examples{

print_minimal_reproducible_code_in_case_of_MRMC()


}
