% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_predictor.R
\name{prior_predictor}
\alias{prior_predictor}
\title{Predict some estimates of parameter}
\usage{
prior_predictor(d = d)
}
\arguments{
\item{d}{A list of data, which can be passed to the \code{fit_Bayesian_FROC}.}
}
\value{
none
}
\description{
Predict some estimates of parameter
}
