% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_print.R
\name{prior_print_MRMC}
\alias{prior_print_MRMC}
\title{Print What Prior Are Used}
\usage{
prior_print_MRMC(prior = 0)
}
\arguments{
\item{prior}{An integer, representing type of Prior}
}
\value{
none
}
\description{
Prints prior in R console
}
\examples{

   prior_print_MRMC()

}
