% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_hat_max.R
\name{R_hat_max}
\alias{R_hat_max}
\title{Max R hat}
\usage{
R_hat_max(StanS4class)
}
\arguments{
\item{StanS4class}{A stanfit object.}
}
\value{
A real number, indicating the maximal R hat over all parameters.
}
\description{
Max R hat
}
