% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Based_Calibration.R
\name{rank_statistics_with_two_parameters}
\alias{rank_statistics_with_two_parameters}
\title{Rank Statistics}
\usage{
rank_statistics_with_two_parameters(
  values.of.f.at.one.MCMC.samples,
  values.of.f.at.a.sample.from.priors
)
}
\arguments{
\item{values.of.f.at.one.MCMC.samples}{The value of f at a vector whose components are constructed by the all parameters at one MCMC sample.}

\item{values.of.f.at.a.sample.from.priors}{The value of f at a vector of model parameters from the prior distribution.}
}
\value{
The value of the Rank Statistics
}
\description{
Rank Statistics
}
\examples{

 \dontrun{
#======== The first example   ==========================================

  rank_statistics_with_two_parameters(c(1,2,3,4,5),4)

#=======  The Second Example  ==========================================

       a <- Draw_a_simulated_data_set_and_Draw_posterior_samples()

    rank_statistics_with_two_parameters(
         a$MCMC.samples.sended.by.fun,
         a$prior.samples.sended.by.fun
         )


}# dottest

}
