% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC.R
\name{ROC_data_creator}
\alias{ROC_data_creator}
\title{Synthesize ROC data}
\usage{
ROC_data_creator(
  Number_of_cases = 100,
  Number_of_non_cases = 100,
  prob_case = c(0.1, 0.2, 0.3, 0.5),
  prob_non_case = c(0.4, 0.3, 0.2, 0.1)
)
}
\arguments{
\item{Number_of_cases}{Number_of_cases}

\item{Number_of_non_cases}{Number_of_non_cases}

\item{prob_case}{prob_case}

\item{prob_non_case}{prob_non_case}
}
\value{
A list, indicatin ROC data
}
\description{
Synthesize ROC data
}
