% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc_new.R
\name{sbcc}
\alias{sbcc}
\title{SBC}
\usage{
sbcc(stanmodel, data, M, iter, refresh)
}
\arguments{
\item{stanmodel}{see \code{?sbc}}

\item{data}{To specify priors.}

\item{M}{The number of samples for rank statistics}

\item{iter}{MCMC iterations}

\item{refresh}{????}
}
\value{
????
}
\description{
Priors should guarante suitable conditions such that the ...
}
\examples{
\dontrun{

stanModel <- stan_model_of_sbc()

Simulation_Based_Calibration_single_reader_single_modality_via_rstan_sbc(
  ite = 233,
  M = 11,
  epsilon = 0.04,
  stanModel = stanModel
)
 }# dontrun


}
\author{
Some Stan developer, I am not sure,..., who?
}
