% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc_new.R
\name{seq_array_ind}
\alias{seq_array_ind}
\title{Makes a Matrix from a vector of itegers}
\usage{
seq_array_ind(d, col_major = FALSE)
}
\arguments{
\item{d}{A vector of integers}

\item{col_major}{A logical, whether,.... ?}
}
\value{
A matrix, dimension is \code{prod(d)} times \code{ length(d)}.
}
\description{
To make sbc funtion
}
\examples{

  a <- seq_array_ind(1:3,col_major = TRUE)
#> a
#
#       [,1] [,2] [,3]
# [1,]    1    1    1
# [2,]    1    2    1
# [3,]    1    1    2
# [4,]    1    2    2
# [5,]    1    1    3
# [6,]    1    2    3






b<-seq_array_ind(1:3,col_major = FALSE)


}
\author{
Some Stan developer, I am not sure,..., who?
}
