% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_codes_in_my_manuscript.R
\name{show_codes_in_my_manuscript}
\alias{show_codes_in_my_manuscript}
\title{Show \R codes used in my manuscript}
\usage{
show_codes_in_my_manuscript()
}
\value{
\code{NULL}
}
\description{
Show \R codes used in my manuscript
}
\examples{
#========================================================================================
#           R codes in my manuscript
#========================================================================================

show_codes_in_my_manuscript()

}
