% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showGraphicalModel.R
\name{showGM}
\alias{showGM}
\title{Show the Graphical Model for a single reader and single modality}
\usage{
showGM()
}
\description{
This function shows the graphical model for a single reader and single modality FROC statistical model.
}
\examples{

  showGM()


\donttest{
}# dottest
}
