% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_of_return_value.R
\name{size_of_return_value}
\alias{size_of_return_value}
\title{Size of R object}
\usage{
size_of_return_value(object, summary = TRUE, is_return_value = TRUE,
  base_size = 0, col = FALSE)
}
\arguments{
\item{object}{Any R object, whose size is measured.}

\item{summary}{A logical, whether the result is printed.}

\item{is_return_value}{A logical, printed word is used as " return value " if it is \code{TRUE}.}

\item{base_size}{This value is added to the return value, namely,  object size + \code{base_size} is the return value. This is for the package developer.}

\item{col}{A logical, wheter print is colored.}
}
\value{
return value of \code{utils::object.size()}
}
\description{
This return value can add each other or any number by  the manner: return + number of \R object
}
