% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_margin.R
\name{small_margin}
\alias{small_margin}
\title{Margin}
\usage{
small_margin(
  Down.oma = 1,
  Left.oma = 1,
  Top.oma = 1,
  Right.oma = 1,
  Down.mar = 1,
  Left.mar = 1,
  Top.mar = 1,
  Right.mar = 1
)
}
\arguments{
\item{Down.oma}{smaller gives larger plot region}

\item{Left.oma}{smaller gives larger plot region}

\item{Top.oma}{smaller gives larger plot region}

\item{Right.oma}{smaller gives larger plot region}

\item{Down.mar}{smaller gives larger plot region}

\item{Left.mar}{smaller gives larger plot region}

\item{Top.mar}{smaller gives larger plot region}

\item{Right.mar}{smaller gives larger plot region}
}
\value{
NONE
}
\description{
If each variable is smaller, then the margin of it is smaller, so plot region become larger.
But title and x axis title will be vanished.
}
\details{
To show FROC curve or signal and noise distributions in Shiny Graphical devices,
the author write down this function \code{small_margin}. By taking margin too small, we gives more larger plot regions
in Shiny Graphicl devices. 2019 August 6
}
\examples{

small_margin()
graphics::plot(1:3,1:3)


small_margin(2,2,2,2)
graphics::plot(1:3,1:3)

small_margin(2,2,2,2,4,4,4,4)
graphics::plot(1:3,1:3)
colors()
graphics::rect(
par()$usr[1],
par()$usr[2],
par()$usr[3],
par()$usr[4],

 col = "steelblue3",
 border = NA)

small_margin(0.1,0.1)
graphics::plot(1:2,1:2, type="n")



}
\seealso{
\code{\link{draw_latent_signal_distribution}()}

\code{\link{draw_latent_noise_distribution}()}

\code{\link{DrawCurves}()}

\code{\link{DrawCurves_srsc}()}
}
