% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippet_for_BayesianFROC.R
\name{snippet_for_BayesianFROC}
\alias{snippet_for_BayesianFROC}
\title{Edit Snippet}
\usage{
snippet_for_BayesianFROC()
}
\value{
nothing
}
\description{
Snippet for the package BayesianFROC. Copy and paste to the snippet edition tools in your R studio for the conforable usage of the package BayesianFROC. This is under construction. To edit snippet, you should open the editor located in Tools > Global options > Code > Edit snippets.
}
\examples{

 snippet_for_BayesianFROC()

}
