% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_priors.R
\name{specify_priors}
\alias{specify_priors}
\title{Define priors}
\usage{
specify_priors()
}
\value{
numerical vector indicating the characteristics of the priors.
}
\description{
Specification of the support for priors
}
