% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/validation_MRMC_Create_dataList_MRMC_Hit_from_rate_etc.R
\name{Stan_code_validation}
\alias{Stan_code_validation}
\title{stan code}
\usage{
Stan_code_validation(z = BayesianFROC::z, mu = BayesianFROC::mu,
  v = BayesianFROC::v, T.or.F = T)
}
\arguments{
\item{z}{thresholds}

\item{mu}{mean}

\item{v}{standard deviation}

\item{T.or.F}{logical, if true hten a logical is return
hit rate <1  and if false hit rate is returned.}
}
\description{
Title
}
\examples{

Stan_code_validation(z=c(4.7,5,6),mu+555,v/1000000000)


Stan_code_validation(z=c(4.7,5,6),mu+5,v/10,T.or.F = FALSE)

#ppp[1,3,4]/denoo[1,3,4]

}
