% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbcVer2.R
\name{stan_model_of_sbc}
\alias{stan_model_of_sbc}
\title{Creates an object of class stanfit of SBC}
\usage{
stan_model_of_sbc(model_ver = 2)
}
\arguments{
\item{model_ver}{An integer, indicating priors}
}
\value{
An object of class stanfit for SBC
}
\description{
Creates an object of class stanfit of SBC
}
\examples{
\dontrun{
stan_model_of_sbc()
}
}
\seealso{
\code{\link{Simulation_Based_Calibration_single_reader_single_modality_via_rstan_sbc}()}
}
