% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argMax.R
\name{stan_trace_of_max_rhat}
\alias{stan_trace_of_max_rhat}
\title{a trace plot for a paramter whose R hat is largest}
\usage{
stan_trace_of_max_rhat(StanS4class)
}
\arguments{
\item{StanS4class}{An S4 object of class \emph{\code{ \link{stanfitExtended}}} which is an inherited class from the S4 class  \code{stanfit}.
This \R object is a fitted model object
 as a return value of the function \code{\link{fit_Bayesian_FROC}()}.

To be passed to \code{\link{DrawCurves}()}, \code{\link{ppp}()}  and ... etc}
}
\value{
none
}
\description{
a trace plot for a paramter whose R hat is largest
}
\examples{
\dontrun{
#========================================================================================
#                   Draw   a trace plot for a paramter whose R hat is largest
#========================================================================================


#  Fit a model to data
#______________________



      f <- fit_Bayesian_FROC(
                         ite  = 111,
                          cha = 1,
                     dataList = d)


# Extract a name of parameter whose R hat is maximal over all parameters
#____________________________________________________________________



 stan_trace_of_max_rhat(f)


}

}
