% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argMax.R
\name{stanfit_from_its_inherited_class}
\alias{stanfit_from_its_inherited_class}
\title{Chage S4 class to stanfit}
\usage{
stanfit_from_its_inherited_class(StanS4class)
}
\arguments{
\item{StanS4class}{An S4 object of class \emph{\code{ \link{stanfitExtended}}} which is an inherited class from the S4 class  \code{stanfit}.
This \R object is a fitted model object
 as a return value of the function \code{\link{fit_Bayesian_FROC}()}.

To be passed to \code{\link{DrawCurves}()} ... etc}
}
\value{
A fitted model object whose S4 class is the stanfit
}
\description{
Chage S4 class to stanfit
}
\examples{
\dontrun{
#========================================================================================
#                   Draw   a trace plot for a paramter whose R hat is largest
#========================================================================================


#  Fit a model to data
#______________________



      f <- fit_Bayesian_FROC(
                         ite  = 111,
                          cha = 1,
                     dataList = d)

# Change the class of the above object "f" to stanfit from its inherited class
#____________________________________________________________________



  stanfit_from_its_inherited_class(f)


}

}
