% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticForANOVA.R
\name{StatisticForANOVA}
\alias{StatisticForANOVA}
\title{Statistic for ANOVA}
\usage{
StatisticForANOVA()
}
\value{
None
}
\description{
Provides a statistic to test the null hypothesis that
all modalities are same.
}
