% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_MRMC.R
\name{summarize_MRMC}
\alias{summarize_MRMC}
\title{Summarize the estimates for MRMC case}
\usage{
summarize_MRMC(StanS4class, dig = 3)
}
\arguments{
\item{StanS4class}{An \R  object of class \emph{\code{ \link{stanfitExtended}}} inherited from the S4 class  \code{\link[rstan]{stanfit}}  that can be passed to the \code{\link{DrawCurves}()} and  \code{\link{p_value_of_the_Bayesian_sense_for_chi_square_goodness_of_fit}()}, ..., etc.}

\item{dig}{To be passed to the function \code{rstan::}\code{\link[rstan]{sampling}}() in \strong{rstan}. An argument of \code{rstan::}\code{\link[rstan]{sampling}}()  in which it is named \code{...??}.   A positive integer representing   the Significant digits, used in stan Cancellation.
default = 5,}
}
\value{
Nothing
}
\description{
Summarize the estimates for MRMC case
}
\examples{
\donttest{
   fit <- fit_Bayesian_FROC(
           dataList.Chakra.Web.orderd,
           ite = 1111,
           summary =FALSE
           )

   summarize_MRMC(fit)


}# dottest
}
\author{
Issei Tsunoda
}
