% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_AUC_comparison_MRMC.R
\name{summary_AUC_comparison_MRMC}
\alias{summary_AUC_comparison_MRMC}
\title{Print summary for AUC comparisons for MRMC}
\usage{
summary_AUC_comparison_MRMC(StanS4class, significantLevel = 0.8,
  dig = 3)
}
\arguments{
\item{StanS4class}{This is an output of \code{rstan::stan()}. More precisely, an object of the inherited class of the S4 class stanfit.
That is, the stanfitExtended S4 class defined in this package.}

\item{significantLevel}{This is a number between 0 and 1. The results are shown if posterior probabilities are greater than this quantity.

 If your model could not converge,
then change this number.}

\item{dig}{An argument of \code{rstan::stan()} in which it is named \code{iter}.  It means the Significant digits, used in stan Cancellation.
default = 5,}
}
\description{
This is print the results of AUC comparison for MRMC data.
}
