% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_AUC_comparison_MRMC_without_crayon.R
\name{summary_AUC_comparison_MRMC_without_crayon}
\alias{summary_AUC_comparison_MRMC_without_crayon}
\title{Print summary for AUC comparisons for MRMC without color}
\usage{
summary_AUC_comparison_MRMC_without_crayon(StanS4class,
  significantLevel = 0.8, dig = 3)
}
\arguments{
\item{StanS4class}{An \R  object of class \emph{\code{ \link{stanfitExtended}}} inherited from the S4 class  \code{\link[rstan]{stanfit}}  that can be passed to the \code{\link{DrawCurves}()} and  \code{\link{p_value_of_the_Bayesian_sense_for_chi_square_goodness_of_fit}()}, ..., etc.}

\item{significantLevel}{This is a number between 0 and 1. The results are shown if posterior probabilities are greater than this quantity.}

\item{dig}{To be passed to the function \code{rstan::}\code{\link[rstan]{sampling}}() in \strong{rstan}. An argument of \code{rstan::}\code{\link[rstan]{sampling}}()  in which it is named \code{...??}.   A positive integer representing   the Significant digits, used in stan Cancellation.
default = 5,}
}
\description{
This is print the results of AUC comparison for MRMC data.
}
