% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Test_Null_Hypothesis_that_all_modalities_are_same.R
\name{Test_Null_Hypothesis_that_all_modalities_are_same}
\alias{Test_Null_Hypothesis_that_all_modalities_are_same}
\title{Test the Null hypothesis that all modalities are same}
\usage{
Test_Null_Hypothesis_that_all_modalities_are_same(dataList, ite = 111,
  cha = 1, summary = FALSE)
}
\arguments{
\item{dataList}{MRMC case only.}

\item{ite}{An argument of \code{rstan::}\code{\link[rstan]{sampling}}()  in which it is named \code{iter}. A positive integer representing  the  number of samples generated by Hamiltonian Monte Carlo method,
and, default = 10000. If your model could not converge, then raise this number. Must be greater for more reliable estimates.}

\item{cha}{An argument of \code{rstan::}\code{\link[rstan]{sampling}}()  in which it is named \code{chains}.  A positive integer representing   the number of chains generated by Hamiltonian Monte Carlo method,
and, default = 1.}

\item{summary}{Logical: \code{TRUE} of \code{FALSE}. Whether to print the verbose summary, i.e., logical; whether to list verbose data of estimates when printing. If TRUE then verbose summary and additional estimates are printed in the R console. If FALSE, the output is minimal. I regret, this variable name should be verbose.}
}
\value{
none
}
\description{
Test null hypothesis that all modalities have same observer performance ablity,
 using Bayes factor.

 @details From input data \code{dataList}, the two objects of class  \strong{\emph{\code{\link[rstan]{stanfit}}}}
 are created. one is fitted to the null hypothesis model and the another one representing alternative hypothesis.
 These two  \strong{\emph{\code{\link[rstan]{stanfit}}}}. objects are compared based on Bayes factor.
}
\examples{
\donttest{

 Test_Null_Hypothesis_that_all_modalities_are_same(BayesianFROC::dd)

 }#donttest
}
