% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_row_number_of_logical_vector.R
\name{the_row_number_of_logical_vector}
\alias{the_row_number_of_logical_vector}
\title{Extract the row number from a logical vector}
\usage{
the_row_number_of_logical_vector(vector.logical)
}
\arguments{
\item{vector.logical}{vector with logical component}
}
\value{
the row number of logical component
}
\description{
Extract the row number from a logical vector
}
\examples{


 a <-c(TRUE,FALSE,FALSE,TRUE,TRUE)


 b <-  the_row_number_of_logical_vector(a)

# Then, return value object, b is a vector of

#> b
#  1, 4, 5

# From this, we can count the TRUE, as following manner:

 Number.of.TRUE <- length(b)

# Of course, it is:
#> Number.of.TRUE
#  3

}
\author{
Issei Tsunoda
}
