% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_row_number_of_logical_vector.R
\name{TRUE.Counter.in.vector}
\alias{TRUE.Counter.in.vector}
\title{Count \code{TRUE} in Logical Vector}
\usage{
TRUE.Counter.in.vector(vector.logical)
}
\arguments{
\item{vector.logical}{vector with logical component}
}
\value{
A number.
}
\description{
For the posterior predictive p value.
}
\examples{

 a <-c(TRUE,FALSE,FALSE,TRUE,TRUE)

 TRUE.Counter.in.vector(a)

# Of course, it is:
#> Number.of.TRUE
#  3

}
