% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_true_param.R
\docType{data}
\name{v_truth}
\alias{v_truth}
\title{Standard Deviation: parameter of an MRMC model}
\description{
A posterior mean of the model parameter for data \code{\link{ddd}} as an example of truth parameter.
}
\details{
Standard Deviation Rate data of some MRMC data to use as a default value
of the function \code{ hits_creator_from_rate}.
This is an array obtained from estimates of some data contained in this package.
To simulate a replication of dataset, the default values should be used from an actual values.
 Thus the author prepare this data.
}
\seealso{
\code{hits_creator_from_rate}
}
\author{
Issei Tsunoda \email{tsunoda.issei1111@gmail.com }
}
