% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/validation_MRMC_Create_dataList_MRMC_Hit_from_rate_etc.R
\name{v_truth_creator_for_many_readers_MRMC_data}
\alias{v_truth_creator_for_many_readers_MRMC_data}
\title{v of MRMC model paramter}
\usage{
v_truth_creator_for_many_readers_MRMC_data(M, Q)
}
\arguments{
\item{M}{An integer, indicating a number of modalities}

\item{Q}{An integer, indicating a number of readers}
}
\value{
An array, representing v of MRMC model paramter
}
\description{
v of MRMC model paramter
}
\examples{

 v <- v_truth_creator_for_many_readers_MRMC_data(M=4,Q=50)


}
