% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_error_srsc.R
\name{validation.draw_srsc}
\alias{validation.draw_srsc}
\title{Draw Curves for validation dataset}
\usage{
validation.draw_srsc(
  validation.data,
  mesh.for.drawing.curve = 11111,
  upper_y = 1,
  DrawFROCcurve = TRUE
)
}
\arguments{
\item{validation.data}{This is a return value of
the function \code{validation.dataset_srsc}.}

\item{mesh.for.drawing.curve}{A positive large integer, indicating number of dots drawing the curves, Default =10000.}

\item{upper_y}{This is a upper bound for the axis of the vertical coordinate of FROC curve.}

\item{DrawFROCcurve}{Logical: \code{TRUE} of \code{FALSE}.  Whether or not FROC curves are shown.}
}
\value{
\code{NULL}
}
\description{
drawing curves.


\strong{Red curve} indicates an FROC curve of truth parameter.

\strong{Other curves} are drawn using replicated estimates.
}
\examples{

\donttest{
#--------------------------------------------------------------------------------------
#                         1)       Draw the curve for each replicated dataset
#--------------------------------------------------------------------------------------

    datasets <- validation.dataset_srsc()

    validation.draw_srsc(datasets)




#--------------------------------------------------------------------------------------
#                         1)       Draw the curve for each replicated dataset
#--------------------------------------------------------------------------------------



              datasets <- validation.dataset_srsc(replicate.datset = 5)


                                 validation.draw_srsc(datasets)


}# dottest
}
