% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertical_to_horizontal.R
\name{vertical_from_horizontal_in_each_case}
\alias{vertical_from_horizontal_in_each_case}
\title{Transfer From Horizontal placement into Vertical  placement for casewise vectors}
\usage{
vertical_from_horizontal_in_each_case(array, NI, M, Q, C)
}
\arguments{
\item{array}{a array, such as hit array, hhh or fff counted for each cases.}

\item{NI}{Number of images, in other words, cases}

\item{M}{Number of modalities}

\item{Q}{Number of readers}

\item{C}{Number of confidence levels, in other words, ratings}
}
\value{
an array whose dimension is NI, M,Q,C
}
\description{
Transfer From Horizontal placement into Vertical  placement for casewise vectors
}
