% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewdata.R
\name{viewdata_MRMC}
\alias{viewdata_MRMC}
\title{View MRMC data}
\usage{
viewdata_MRMC(dataList, summary = TRUE, head.only = FALSE)
}
\arguments{
\item{dataList}{it should include  \code{m,q,c,h,f,NL,C,M,Q} which means from the right

\code{m } means the modality ID vector

\code{q } means the reader ID vector

\code{c } means the confidence level

\code{h } means the number of hits

\code{f } means the number of false alarm

\code{NL } means the Total number of lesions for all images

\code{C } means the highest number of confidence level

\code{M } means the number of modalities

\code{Q } means the number of readers.

The detail of these dataset, please see the endowed datasets.
Note that the maximal number of confidence  level, denoted by  \code{C}, are included,
however,
its each confidence level should not included  your data. So, to confirm
your false positives and hits are correctly correspondence
to confidence levels,
you should confirm the orders by the function \code{viewdata_MRMC}.}

\item{summary}{TRUE or FALSE, if true then results are printed, if FALSE this function do nothing.}

\item{head.only}{Logical: \code{TRUE} of \code{FALSE}. Whether  head part or entire. If \code{TRUE}, only head part are shown. Default is \code{FALSE}}
}
\description{
Build a table for data \code{dataList}.
}
