% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewdata_srsc.R
\name{viewdata_srsc}
\alias{viewdata_srsc}
\title{Build a table of data in the case of
Single reader and Single modality (srsc)}
\usage{
viewdata_srsc(dataList, summary = TRUE)
}
\arguments{
\item{dataList}{it should include  \code{f, h, NL, NI, C}.
The detail of these dataset, please see the endowed datasets.
Note that the maximal number of confidence  level, denoted by  \code{C}, are included,
however,
its each confidence level should not included  your data. So, to confirm
your false positives and hits are correctly correspondence
to confidence levels,
user should confirm the orders by the function.}

\item{summary}{TRUE or FALSE, if true then results are printed, if FALSE this function do nothing.}
}
\description{
In order to confirm your data, please view table.
my program makes new column of confidence levels which are used in
my program. So, it is possible that your order of confidence level and
Program's order of confidence level are inverse.
This function's result table are the one which are used in program.
}
\examples{
\donttest{
#   First, we prepare an example FROC data "dataList.Chakra.1" in this package.
# Note that this data should be formed as a single reader and single case (modality).
# If your data are a multiple reader and multiple case (modality), i.e.,MRMC-data,
# then another function named viewdataMRMC is available for MRMC-data.

             dat  <- get(data("dataList.Chakra.1"))




# Show data named "dat";


             viewdata_srsc(dat)



#The Reason why the author made this \\code{viewdata_srsc} is
#the code does not refer your confidence level.
#More precisely, my program made the column vector of confidence levels
#from the its highest number,
#so, it may be occur the interpretaion of code for hits and false alarm
#are inverse order compared with your data.


}# dottest

}
