% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_true_param.R
\docType{data}
\name{z}
\alias{z}
\title{Threshold: parameter of an MRMC model}
\description{
A posterior mean of the model parameter for data \code{\link{ddd}} as an example of truth parameter.
}
\seealso{
\code{make_true_parameter_MRMC}
}
\author{
Issei Tsunoda \email{tsunoda.issei1111@gmail.com }
}
