% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFROC.R
\name{AFROC}
\alias{AFROC}
\title{AFROC curve}
\usage{
AFROC(x, a = 0.13, b = 0.19, x.coordinate.also = FALSE)
}
\arguments{
\item{x}{A vector, for which AFROC curve is calculateld.}

\item{a}{A number, for which AFROC curve is calculateld.}

\item{b}{A number, for which AFROC curve is calculateld.}

\item{x.coordinate.also}{Logical, whether a vector of \code{1-exp(-x)}
is included in a return value.}
}
\value{
A list, contains two vectors of x,y cooridinates
 for drawing curves. Or A vector, of y coodinate only,
  (x coodinates is omitted.)
}
\description{
AFROC curve
}
\examples{


     AFROC(stats::runif(1000,3,100),x.coordinate.also=TRUE)



}
