% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FROC_via_ggplot.R
\name{AFROC_curve}
\alias{AFROC_curve}
\title{FROC curve as an embedding map}
\usage{
AFROC_curve(x)
}
\arguments{
\item{x}{A real number moves in domain  of FROC curve}
}
\value{
none
}
\description{
FROC curve as an embedding map
}
\details{
AFROC is difficult since it has two infiniy gradient points and it cause the error.
Warning messages:
1: In stats::qnorm(exp(1 - x)) : NaNs produced
2: In stats::qnorm(exp(1 - x)) : NaNs produced
3: Removed 50 rows containing missing values (geom_path).
}
\examples{

# I love you!



}
