% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Author_vs_Chakraborty_for_AUC.R
\name{Author_vs_classic_for_AUC}
\alias{Author_vs_classic_for_AUC}
\title{validation of AUC calculation}
\usage{
Author_vs_classic_for_AUC(StanS4class)
}
\arguments{
\item{StanS4class}{An S4 object of class \emph{\code{ \link{stanfitExtended}}} which is an inherited class from the S4 class  \code{\link[rstan]{stanfit}}.
This \R object is a fitted model object
 as a return value of the function \code{\link{fit_Bayesian_FROC}()}.

It can be passed to \code{\link{DrawCurves}()}, \code{\link{ppp}()}  and ... etc}
}
\value{
AUCs
}
\description{
This is for the author.
}
\author{
Issei Tsunoda
}
