% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rhat.R
\name{check_rhat}
\alias{check_rhat}
\title{Diagnosis for convergence}
\usage{
check_rhat(StanS4class)
}
\arguments{
\item{StanS4class}{This is a return value of the function \code{rstan::stan()}, i.e., an object of the S4 class the so-called stanfit.}
}
\value{
TRUE of FALSE. If model converges then TRUE, and if not FALSE.
}
\description{
This function evaluate your model's R hat statistics.
}
\references{
Gelman A. \&  Rubin, D.B. (1992). Inference from Iterative Simulation Using Multiple Sequences, Statistical Science, Volume 7, Number 4, 457-472.
}
\author{
betanalpha. This is not my function. But I modified it.
}
