% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset.R
\name{create_dataset}
\alias{create_dataset}
\title{Creates a dataset}
\usage{
create_dataset()
}
\value{
A list of FROC data to which we fit a FROC model.

2019 Dec 12
}
\description{
Creates a dataset to apply the  function \code{ \link{fit_Bayesian_FROC} }.
}
\details{
This is an interactive creator of an FROC dataset.
    Using this return value, we
     can fit a FROC model to data
     by applying the function
      \code{ \link{fit_Bayesian_FROC}} in this package.

      To tell the truth, the author never use this function to create datset.
      So,... this function is not so good.
}
\examples{

\dontrun{
## Only run examples in interactive R sessions
if (interactive()) {


     create_dataset()


}### Only run examples in interactive R sessions
}

}
