% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_dataset.R
\docType{data}
\name{d}
\alias{d}
\title{Single reader and Single modality data}
\description{
This is used to build a hierarchical FROC model. This data is same as dataList.Chakra.1.
}
\details{
This data is same as \code{\link{dataList.Chakra.1.with.explantation}}.
 The author name it  \code{d} for the sake of simplicity.
}
\references{
Maximum likelihood analysis of free-response  receiver operating characteristic (FROC) data, Dev P. Chakraborty.
}
\seealso{
\code{\link{dataList.Chakra.1.with.explantation}}
}
\author{
Issei Tsunoda \email{tsunoda.issei1111@gmail.com }
}
