% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dark_theme.R
\name{dark_theme}
\alias{dark_theme}
\title{Dark Theme}
\usage{
dark_theme(type = 1)
}
\arguments{
\item{type}{1,2,3,4,5,6,7,...}
}
\value{
Nothing
}
\description{
Putting it before plotting, plot area become the dark theme.
}
\details{
a function which indicates what color are used when we use graphic devices.
}
\examples{
 \donttest{

 dark_theme(1)

 graphics::plot(c(1,2,3),c(1,2,3))


 dark_theme(2)

 graphics::plot(c(1,2,3),c(1,2,3))


 # 2019.05.21 Revised.

}# dottest
}
