% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_dataset_MRMC.R
\docType{data}
\name{data.hier.ficitious}
\alias{data.hier.ficitious}
\title{Multiple reader and Multiple modality data}
\description{
This is used to build a hierarchical FROC model.
}
\details{
This data is fictitious.
}
\references{
The author' preprint
}
\author{
Issei Tsunoda \email{tsunoda.issei1111@gmail.com }
}
