% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_dataset.R
\docType{data}
\name{dataList.Chakra.1}
\alias{dataList.Chakra.1}
\title{Single reader and Single modality data}
\description{
This is used to build a non-hierarchical FROC model.
}
\details{
This data appeared in Chakraborty's paper (1988)
}
\references{
Maximum likelihood analysis of free-response receiver operating characteristic (FROC) data, Dev P. Chakraborty.
}
\seealso{
\code{\link{dataList.Chakra.1.with.explantation}}
}
\author{
Issei Tsunoda \email{tsunoda.issei1111@gmail.com }
}
