% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_creator_new_version.R
\name{dataset_creator_new_version}
\alias{dataset_creator_new_version}
\title{Create a Dataset (version 2)  Interactively}
\usage{
dataset_creator_new_version()
}
\value{
A  list  representing FROC data,
to build FROC fitted model
object by \code{\link{fit_Bayesian_FROC}()}.
}
\description{
Create the Passing data to
 the  function \code{ \link{fit_Bayesian_FROC}}.

     This is an interactive creator of dataset for FROC data.
}
\details{
This provide the intaractive making
of FROC dataset by using table to
 summarize hits and false alarm data.


Using this return value, you can build
the FROC model for your data by applying
the function  \code{ \link{fit_Bayesian_FROC}()  }
in this package.


Should carefully for the order of confidence levels.
}
