% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_Bayesian_FROC.R
\name{demo_Bayesian_FROC}
\alias{demo_Bayesian_FROC}
\title{demonstration}
\usage{
demo_Bayesian_FROC()
}
\value{
none
}
\description{
demonstration
}
\details{
The author often forget the R script for execute the demos or bother to write the code to execute demo, thus I made this.
}
\examples{
 \donttest{


 demo_Bayesian_FROC()

# 2019.05.21 Revised.
}# dottest
}
