% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_Bayesian_FROC_without_pause.R
\name{demo_Bayesian_FROC_without_pause}
\alias{demo_Bayesian_FROC_without_pause}
\title{demonstration without pause}
\usage{
demo_Bayesian_FROC_without_pause()
}
\value{
none
}
\description{
demonstration without pause
}
