% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draw_an_area_of_AUC_for_srsc.R
\name{Draw_an_area_of_AUC_for_srsc}
\alias{Draw_an_area_of_AUC_for_srsc}
\title{Draw a Region of the area under the AFROC curve}
\usage{
Draw_an_area_of_AUC_for_srsc(StanS4class)
}
\arguments{
\item{StanS4class}{An S4 object of class \emph{\code{ \link{stanfitExtended}}} which is an inherited class from the S4 class  \code{\link[rstan]{stanfit}}. This \R object can be passed to the \code{\link{DrawCurves}()}, \code{\link{ppp}()}  and ... etc}
}
\value{
None
}
\description{
Draw a Region of the area under the AFROC curve
}
\examples{
 \donttest{

 fit <- fit_Bayesian_FROC(dataList.Chakra.1)

 Draw_an_area_of_AUC_for_srsc(fit)



}# dottest
}
