% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draw_an_area_of_AUC_for_srsc.R
\name{Draw_an_area_of_AUC_for_srsc}
\alias{Draw_an_area_of_AUC_for_srsc}
\title{Draw a Region of the area under the AFROC curve}
\usage{
Draw_an_area_of_AUC_for_srsc(StanS4class)
}
\arguments{
\item{StanS4class}{This is an \R  object of class \emph{\code{ \link{stanfitExtended}}} inherited from the S4 class  \strong{\emph{\code{\link[rstan]{stanfit}}}}.}
}
\value{
None
}
\description{
Draw a Region of the area under the AFROC curve
}
\examples{
 \donttest{

 fit <- fit_Bayesian_FROC(dataList.Chakra.1)

 Draw_an_area_of_AUC_for_srsc(fit)



}# dottest
}
