% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawCurves_MRMC.R
\name{DrawCurves_MRMC}
\alias{DrawCurves_MRMC}
\title{Draw the FROC  curves for all modalities and readers}
\usage{
DrawCurves_MRMC(StanS4class)
}
\arguments{
\item{StanS4class}{An \R  object of class \emph{\code{ \link{stanfitExtended}}} inherited from the S4 class  \code{\link[rstan]{stanfit}}  that can be passed to the \code{\link{DrawCurves}()} and  \code{\link{p_value_of_the_Bayesian_sense_for_chi_square_goodness_of_fit}()}, ..., etc.}
}
\description{
Draw the FROC  curves and AFROC curves for all modalities and readers, if many  modalities and readers exists, then so very confused plots will be drawn.
}
\examples{

\donttest{
  fit <- fit_Bayesian_FROC(
                           dataList.Chakra.Web.orderd,
                           ite = 1111,
                           summary =FALSE
                           )

               DrawCurves_MRMC(fit)



# 2019.05.21 Revised.

}# dottest

}
\author{
Issei Tsunoda
}
