% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawCurves_MRMC.R
\name{DrawCurves_MRMC}
\alias{DrawCurves_MRMC}
\title{Draw the FROC  curves for all modalities and readers}
\usage{
DrawCurves_MRMC(StanS4class)
}
\arguments{
\item{StanS4class}{Fitted model object created by \code{fit_Bayesian_FROC()}. The class of this object is stanfitExtended and not the class stanfit.}
}
\description{
Draw the FROC  curves and AFROC curves for all modalities and readers, if many  modalities and readers exists, then so very confused plots will be drawn.
}
\examples{

\donttest{
  fit <- fit_Bayesian_FROC(
                           dataList.Chakra.Web.orderd,
                           ite = 1111,
                           summary =FALSE
                           )

               DrawCurves_MRMC(fit)



# 2019.05.21 Revised.

}# dottest

}
