% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawCurves.R
\name{DrawCurves_MRMC}
\alias{DrawCurves_MRMC}
\title{Draw the FROC
 curves for all modalities and readers}
\usage{
DrawCurves_MRMC(StanS4class, type = 1)
}
\arguments{
\item{StanS4class}{An S4 object of class \emph{\code{ \link{stanfitExtended}}} which is an inherited class from the S4 class  \code{stanfit}.
This \R object is a fitted model object
 as a return value of the function \code{\link{fit_Bayesian_FROC}()}.

It can be passed to \code{\link{DrawCurves}()}, \code{\link{ppp}()}  and ... etc}

\item{type}{An integer, for the color of background and etc.}
}
\description{
Draw the FROC
 curves and AFROC curves for all
 specified modalities and readers.
}
\examples{

\dontrun{
  fit <- fit_Bayesian_FROC(
                           dataList.Chakra.Web.orderd,
                           ite = 1111,
                           summary =FALSE
                           )

               DrawCurves_MRMC(fit)



# 2019.05.21 Revised.

}# dottest

}
