% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_message.R
\name{error_message}
\alias{error_message}
\title{Error Message for Data Format}
\usage{
error_message(h, NL)
}
\arguments{
\item{h}{A non-negative integer vector}

\item{NL}{A positive number, indicating Number of lesions}
}
\value{
Plot of error message by the generic function \code{plot()}. So, return value is not required.
}
\description{
This function is excellent! I fear my own great genius. How great I am ,... Is there someone?
Who read this? I always feel vanity when write this mannual, who read? ha,.... This function is only reture plot to let user know the data format is error.
}
\details{
Why the author use the generic funtion \code{plot} instead of such as \code{message()} or \code{cat()} is for \emph{Shiny}.
So, this error message is shown in the plot plane in the Graphical User Interface, so in there,  \code{message()} or \code{cat()} cannot use.
Ha,..who read? I feel empty. In mathematics empty set is very vain. My heart is now, empty set. ha,,, I love you.
}
\examples{
#----------------------------------------------------------------------------------------
#            If   number of hits > number of lesion,  then error message appears.
#----------------------------------------------------------------------------------------

 # Make an example such that sum(h) > NL, that is, the sum of the number of hits is
 # greater than the number of lesion, then, it launced error message.

 h <- c(50,30,20)
 NL <- 3


 error_message(h,NL)

 # Then, in imaging device, the error message are shown.
 # In shiny, even if plot cannot be done causing some error, Graphical User Interface
 # can not change, so I have to use the graphical user interface.
 # Thus. in such case, I chose this function rather than the message() or cat().

 # Who read this? My heart will be more empty when I wrote this mannual.
 # Now, today, my health is good, so I want to go to eat Sushi,...ha,  yari_ika_geso.

 # This function is made in 2019 July, 6.

}
\seealso{
\code{ \link{fit_GUI}() }
}
\author{
Issei Tsunoda
}
