% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractAUC.R
\name{extractAUC}
\alias{extractAUC}
\title{Extract AUC}
\usage{
extractAUC(StanS4class, dig = 3, summary = TRUE)
}
\arguments{
\item{StanS4class}{An \R  object of class \emph{\code{ \link{stanfitExtended}}} inherited from the S4 class  \code{\link[rstan]{stanfit}}  that can be passed to the \code{\link{DrawCurves}()} and  \code{\link{p_value_of_the_Bayesian_sense_for_chi_square_goodness_of_fit}()}, ..., etc.}

\item{dig}{digits of estimates.}

\item{summary}{Logical: \code{TRUE} of \code{FALSE}. Whether to print the verbose summary, i.e., logical; If \code{TRUE} then verbose summary is printed in the \R console. If \code{FALSE}, the output is minimal. I regret, this variable name should be verbose.}
}
\value{
The estimates of AUC with respect to modalities.
The MRMC case, return value is NULL !? Why !? 2019.6.13
}
\description{
Extract AUC for both srsc and MRMC data.
}
