% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractAUC.R
\name{extractAUC}
\alias{extractAUC}
\title{Extract AUC for each modality from hierarchical Bayesian models.}
\usage{
extractAUC(StanS4class, dig = 3, summary = TRUE)
}
\arguments{
\item{StanS4class}{This is an output of \code{rstan::stan} for a single reader and a single modality.
More precisely, this is an object of some inherited class from the S4 class called stanfit in the rstan package.}

\item{dig}{digits of estimates.}

\item{summary}{Logical: \code{TRUE} of \code{FALSE}. Whether to print the verbose summary, i.e., logical; whether to list verbose data of estimates when printing. If TRUE then verbose summary and additional estimates are printed in the R console. If FALSE, the output is minimal. I regret, this variable name should be verbose.}
}
\value{
The estimates of AUC with respect to modalities.
}
\description{
EAPs
}
