% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_GUI_MRMC.R
\name{fit_GUI_MRMC}
\alias{fit_GUI_MRMC}
\title{Fit with GUI via Shiny in case of MRMC}
\usage{
fit_GUI_MRMC(M = 2, Q = 3, C = 4)
}
\arguments{
\item{M}{No. of modality}

\item{Q}{No. of reader}

\item{C}{No. of confidence level}
}
\value{
None
}
\description{
First, please execute, then user will understand what it is.
This function is the one of the most important function in this package.
I do not assume the user is familiar with R script but FROC analysis. So, I made this function to provide the Graphical User Interface (GUI) for users.
I hope it helps someone in the world.
}
