% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posterior_variance.R
\name{get_posterior_variance}
\alias{get_posterior_variance}
\title{Alternative of \code{rstan::get_posterior_mean()}}
\usage{
get_posterior_variance(StanS4class, name.of.parameter)
}
\arguments{
\item{StanS4class}{An S4 object of class \emph{\code{ \link{stanfitExtended}}} which is an inherited class from the S4 class  \code{stanfit}.
This \R object is a fitted model object
 as a return value of the function \code{\link{fit_Bayesian_FROC}()}.

To be passed to \code{\link{DrawCurves}()}, \code{\link{ppp}()}  and ... etc}

\item{name.of.parameter}{An parameter name (given as a character string, should not surround by "").
The name of parameter which user want to extract.
Parameters are contained in the parameter block of each Stan file in the path: inst/extdata.}
}
\value{
variance or posterior parameters, if it is an array, then return is also an array.
}
\description{
This function is underconstruction. I validate only the example of this function.
For MRMC case, I have to write or modify code. 2019 Sept 6
}
\examples{



\dontrun{

        fit <- fit_Bayesian_FROC(BayesianFROC::dd,ite = 111)




             e <- rstan::extract(fit)




 # Check the retrun value is the desired one.


#    apply(e$z,   2,        var) ==  get_posterior_variance(fit,z)
#    apply(e$mu,  c(2,3),   var) ==  get_posterior_variance(fit,mu)
#    apply(e$v,   c(2,3),   var) ==  get_posterior_variance(fit,v)
#    apply(e$ppp, c(2,3,4), var) ==  get_posterior_variance(fit,ppp)

#This code is OK, but R CMD check might say error cuz the object
# z, mu, v, ppp is not found



# apply(e$z,   2,        var) ==  get_posterior_variance(fit,"z")
# apply(e$mu,  c(2,3),   var) ==  get_posterior_variance(fit,"mu")
# apply(e$v,   c(2,3),   var) ==  get_posterior_variance(fit,"v")
# apply(e$ppp, c(2,3,4), var) ==  get_posterior_variance(fit,"ppp")


 }#dontrun





}
