% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_name_srsc_data.R
\name{give_name_srsc_CFP_CTP_vector}
\alias{give_name_srsc_CFP_CTP_vector}
\title{Give a Name For CTP CFP vector}
\usage{
give_name_srsc_CFP_CTP_vector(vector, CFP.or.CTP = "CFP",
  ModifiedPoisson = FALSE)
}
\arguments{
\item{vector}{A vector representing cumulative true positives (CTPs) or cumulative false positives (CFPs).}

\item{CFP.or.CTP}{"CFP" or "CTP". Default value is “CFP”.}

\item{ModifiedPoisson}{This is dichotomous, that is \code{TRUE} or \code{FALSE}. If
\code{ModifiedPoisson = TRUE}, then Poisson rate of false alarm are \code{per lesion},
and if \code{ModifiedPoisson = FALSE}, then Poisson rate of false alarm are \code{per image}.
To know detail, refer the author's paper in which I explained per image and per lesion.}
}
\value{
A vector representing cumulative true positives (CTPs) or cumulative false positives (CFPs) with its name.
}
\description{
Give a Name for a vector representing cumulative true positives (CTPs) or cumulative false positives (CFPs).
}
\details{
Some function in this package give the return values of vectors representing the CFP or CTPs. Using this function, we specify what the components of vector means. This is important since its order is not deterministic, that is, its order give two case, one is decreasing and one is increasing order. So, to avoid such confusion, the name should be specified. Of course this function is no needed for user to know or to use it.
}
\examples{

   h <- BayesianFROC::dataList.Chakra.1$h

   NL <- BayesianFROC::dataList.Chakra.1$NL

   CTP.vector <- cumsum(h)/NL

   CTP.vector.with.name <- give_name_srsc_CFP_CTP_vector(CTP.vector)


}
