% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_srsc_per_image.R
\name{metadata_srsc_per_image}
\alias{metadata_srsc_per_image}
\title{Create metadata for MRMC data.}
\usage{
metadata_srsc_per_image(dataList)
}
\arguments{
\item{dataList}{A list, should include  \code{m,q,c,h,f,NL,C,M,Q} which means from the right


\code{c } means the confidence level is not required,
however it is created by \code{  c <-c(rep(C:1))}, where \code{C} is the number of confidence levels.
So, you should write down your hits and false alarms vector so that it is compatible with this automatically created \code{c} vector.

\code{h } means the number of hits

\code{f } means the number of false alarm

\code{NL } means the Total number of lesions for all images

\code{C } means the highest number of confidence level}
}
\value{
A metadata such as number of cumulative false alarms and hits to create and draw the curve.
}
\description{
From data of number of hits and false alarms, we calculate the number of cumulative false positives and hits.
Since there are three subscripts, reader, modality, and image, we create array format and vector format etc...
}
\examples{


\donttest{
#First, we prepare the data endowed with this package.

        dat  <- BayesianFROC::dataList.Chakra.Web




#Second, we run the stan funtion
#with data named "dat"  and the author's Bayesian model.


             metadata_srsc_per_image(dat)






#Now, we get a metadata.
}# dottest
}
\author{
Issei Tsunoda
}
