% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs_plot_if_divergent_transition_occurred.R
\name{pairs_plot_if_divergent_transition_occurred}
\alias{pairs_plot_if_divergent_transition_occurred}
\title{Pairs plot for divergent transition}
\usage{
pairs_plot_if_divergent_transition_occurred(
  StanS4class,
  character.representing.paramter = "z"
)
}
\arguments{
\item{StanS4class}{An S4 object of class \emph{\code{ \link{stanfitExtended}}} which is an inherited class from the S4 class  \code{stanfit}.
This \R object is a fitted model object
 as a return value of the function \code{\link{fit_Bayesian_FROC}()}.

To be passed to \code{\link{DrawCurves}()}, \code{\link{ppp}()}  and ... etc}

\item{character.representing.paramter}{Character, surrounded by "", indicating the paramter of model.}
}
\description{
If divergent transition occurs, the author often forget the variable par or pars. So, I made this to avoid such confusion.
}
\examples{
\dontrun{

# Create a fitted model object of class stanfitExtended  inherited from stanfit.

 fit <- fit_Bayesian_FROC( ite  = 1111,
                           summary = FALSE,
                           cha = 1,
                           Null.Hypothesis = FALSE,
                           dataList = dd )

# Pairs plot to examine the divergent transition.



          # pairs_plot_if_divergent_transition_occurred(fit)

# R CMD check launched error that pkg cannot be found, but it exsits
# Moreover it is available without errors from R console.  but I put # here
# to proceed futher steps in R CMD checks, what a lovely, pretty cute R CMD check is!











    Close_all_graphic_devices()

    }


}
