% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{plot,stanfitExtended,missing-method}
\alias{plot,stanfitExtended,missing-method}
\alias{plot,ANY,ANY-method}
\alias{======}
\title{A  generic function \code{plot()}}
\usage{
\S4method{plot}{stanfitExtended,missing}(x, y, ...)
}
\arguments{
\item{x}{An \R object of the S4 class (\code{\link{stanfitExtended}})}

\item{y}{An \R object of the S4 class   \strong{\emph{\code{\link[methods]{missing-class}}}}.}

\item{...}{Additional arguments}
}
\description{
A  generic function \code{plot()}
}
