% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{plot,stanfitExtended,missing-method}
\alias{plot,stanfitExtended,missing-method}
\title{A method for a generic function plot()}
\usage{
\S4method{plot}{stanfitExtended,missing}(x)
}
\arguments{
\item{x}{This is an object of an S4 class named \strong{\code{ \link{stanfitExtended}}} which is an inherited S4 class from the S4 class \strong{\emph{\code{\link[rstan]{stanfit}}}}  in the rstan package.}
}
\description{
This is a  method \code{print()} for a new  S4 class \strong{\code{ \link{stanfitExtended}}}.
}
\details{
This is a inherited methods from a method of stanfit, plot as an stanfit object and further plot an FROC curve if not MRMC.
}
