% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFROC.R
\name{plotFROC}
\alias{plotFROC}
\title{Draw FROC curves by two parameters a and b}
\usage{
plotFROC(a, b, mesh.for.drawing.curve = 10000, upper_x = 1,
  upper_y = 1, lower_y = 0)
}
\arguments{
\item{a}{An arbitrary real number.
It is no need to require any assumption,
but I use such as \code{a}=\eqn{\mu/\sigma},
where \eqn{\mu} is a mean of signal distribution and \eqn{\sigma} is its standard deviation in the bi-normal assumption.}

\item{b}{An arbitrary positive real number.
I use such as \code{b}=\eqn{1/\sigma},
where \eqn{\sigma} is a standard deviation of signal distribution in the bi-noraml assumption.}

\item{mesh.for.drawing.curve}{An integer indicating number of dots drawing the curves, default =10000.}

\item{upper_x}{The frame size of drawing picture.}

\item{upper_y}{The frame size of drawing picture.}

\item{lower_y}{The frame size of drawing picture.}
}
\description{
Plot FROC curves based on two parameters a and b.
}
