% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{print,stanfitExtended-method}
\alias{print,stanfitExtended-method}
\title{This is a method for a generic function print}
\usage{
\S4method{print}{stanfitExtended}(x)
}
\arguments{
\item{x}{This is an object of an S4 class named stanfitExtended which is an inherited S4 class from the stanfit S4 class in the rstan package.}
}
\description{
This is a method for print and stanfitExtended S4 class.
}
\examples{
 \donttest{

#================The First Example======================================


#(1)First, we prepare the data from this package.

         dat  <- BayesianFROC::dataList.Chakra.1


#(2)Second, we run fit_Bayesian_FROC() in which the rstan::stan() is implemented.
#with data named "dat"  and the author's Bayesian model.


         fit <-  fit_Bayesian_FROC(dat)

#(3)Thirdly, we obtain the return value fit which is an object from an S4 class
# named stanfitExtended that is an inherited class from rstan S4 class stanfit.
# For the S4 class stanfitExtended defined in this package, we can use
# the generic function print for this new S4 class.

       print(fit)



#================The Second Example======================================


#(1)First, we prepare the data from this package.

         dat  <- BayesianFROC::dataList.Chakra.Web


#(2)Second, we run fit_Bayesian_FROC() in which the rstan::stan() is implemented.
#with data named "dat"  and the author's Bayesian model.


         fit <-  fit_Bayesian_FROC(dat)

#(3)Thirdly, we obtain the return value fit which is an object from an S4 class
# named stanfitExtended that is an inherited class from rstan S4 class stanfit.
# For the S4 class stanfitExtended defined in this package, we can use
# the generic function print for this new S4 class.

       print(fit)







# Adding a new method for generic function "print".

}# dottest
}
