% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print_stanfitExtended}
\alias{print_stanfitExtended}
\title{# Definition of a method for the inherited class stanfitExtended from stanfit}
\usage{
print_stanfitExtended(x)
}
\arguments{
\item{x}{This is an object of an S4 class named stanfitExtended which is an inherited S4 class from the stanfit S4 class in the rstan package.}
}
\description{
This is a function for a method in the generic function print.
}
