% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortAUC.R
\name{sortAUC}
\alias{sortAUC}
\title{make a Ranking for AUCs for MRMC data}
\usage{
sortAUC(StanS4class)
}
\arguments{
\item{StanS4class}{An \R  object of class \emph{\code{ \link{stanfitExtended}}} inherited from the S4 class  \code{\link[rstan]{stanfit}}  that can be passed to the \code{\link{DrawCurves}()} and  \code{\link{p_value_of_the_Bayesian_sense_for_chi_square_goodness_of_fit}()}, ..., etc.}
}
\value{
Sorted table of AUC and modality ID
}
\description{
print a modality ranking according to their AUCs.
}
\details{
This is ranking. Sort a datafrmae involving AUC and corresponding modality IDs
}
\examples{

 \donttest{

 fit <- fit_Bayesian_FROC(
 ite  = 1111,
 summary = FALSE,
 cha = 1,
 Null.Hypothesis = FALSE,
  dataList = dd )






 sortAUC(fit)

 }



}
