% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbcVer2.R
\name{stan_model_of_sbc}
\alias{stan_model_of_sbc}
\title{Creates an object of class stanfit of SBC}
\usage{
stan_model_of_sbc()
}
\value{
An object of class stanfit for SBC
}
\description{
Creates an object of class stanfit of SBC
}
\examples{
\dontrun{
stan_model_of_sbc()
}
}
