% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_MRMC.R
\name{summarize_MRMC}
\alias{summarize_MRMC}
\title{Summarize the estimates for MRMC case}
\usage{
summarize_MRMC(StanS4class, dig = 3)
}
\arguments{
\item{StanS4class}{This is an output of \code{rstan::stan()}.}

\item{dig}{An argument of \code{rstan::stan()} in which it is named \code{iter}.  It means the Significant digits, used in stan Cancellation.
default = 5,}
}
\value{
Nothing
}
\description{
Summarize the estimates for MRMC case
}
\examples{
\donttest{
   fit <- fit_Bayesian_FROC(
           dataList.Chakra.Web.orderd,
           ite = 1111,
           summary =FALSE
           )

   summarize_MRMC(fit)


}# dottest
}
