% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_AUC_comparison_MRMC.R
\name{summary_AUC_comparison_MRMC}
\alias{summary_AUC_comparison_MRMC}
\title{Print summary for AUC comparisons for MRMC}
\usage{
summary_AUC_comparison_MRMC(StanS4class, significantLevel = 0.8, dig = 3)
}
\arguments{
\item{StanS4class}{An S4 object of class \emph{\code{ \link{stanfitExtended}}} which is an inherited class from the S4 class  \code{stanfit}.
This \R object is a fitted model object
 as a return value of the function \code{\link{fit_Bayesian_FROC}()}.

It can be passed to \code{\link{DrawCurves}()}, \code{\link{ppp}()}  and ... etc}

\item{significantLevel}{This is a number between 0 and 1. The results are shown if posterior probabilities are greater than this quantity.}

\item{dig}{A variable to be passed to the function \code{rstan::}\code{sampling}() of \pkg{rstan}  in which it is named \code{...??}.   A positive integer representing   the Significant digits, used in stan Cancellation.
Default = 5,}
}
\description{
It prints the results of AUC comparison for MRMC data.
}
