% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_AUC_comparison_MRMC_with_crayon.R
\name{summary_AUC_comparison_MRMC_with_crayon}
\alias{summary_AUC_comparison_MRMC_with_crayon}
\title{Print summary for AUC comparisons for MRMC hier with color}
\usage{
summary_AUC_comparison_MRMC_with_crayon(StanS4class,
  significantLevel = 0.8, dig = 3)
}
\arguments{
\item{StanS4class}{This is an \R  object of class \emph{\code{ \link{stanfitExtended}}} inherited from the S4 class  \strong{\emph{\code{\link[rstan]{stanfit}}}}.}

\item{significantLevel}{This is a number between 0 and 1. The results are shown if posterior probabilities are greater than this quantity.}

\item{dig}{An argument of \code{rstan::}\code{\link[rstan]{sampling}}()  in which it is named \code{iter}.   A positive integer representing   the Significant digits, used in stan Cancellation.
default = 5,}
}
\description{
This is print the results of AUC comparison for MRMC data.
}
