% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Test_Null_Hypothesis_that_all_modalities_are_same.R
\name{Test_Null_Hypothesis_that_all_modalities_are_same}
\alias{Test_Null_Hypothesis_that_all_modalities_are_same}
\title{Test the Null hypothesis that all modalities are same}
\usage{
Test_Null_Hypothesis_that_all_modalities_are_same(
  dataList,
  ite = 1111,
  cha = 1,
  summary = FALSE
)
}
\arguments{
\item{dataList}{MRMC is the  only case in which the function is available for this function.}

\item{ite}{A variable to be passed to the function \code{rstan::}\code{sampling}() of \pkg{rstan}  in which it is named \code{iter}. A positive integer representing  the  number of samples synthesized by Hamiltonian Monte Carlo method,
and, Default = 1111}

\item{cha}{A variable to be passed to the function \code{rstan::}\code{sampling}() of \pkg{rstan}  in which it is named \code{chains}.  A positive integer representing   the number of chains generated by Hamiltonian Monte Carlo method,
and, Default = 1.}

\item{summary}{Logical: \code{TRUE} of \code{FALSE}. Whether to print the verbose summary. If \code{TRUE} then verbose summary is printed in the \R console. If \code{FALSE}, the output is minimal. I regret, this variable name should be verbose.}
}
\value{
none
}
\description{
Test null hypothesis that all modalities
 have same observer performance ablity,
 using Bayes factor.
}
\details{
From input data (variable: \code{dataList}), the two objects
of class  \strong{\emph{\code{stanfit}}}
 are created. one is fitted to the null hypothesis model and
 the another one representing
 alternative hypothesis.
 These two  \strong{\emph{\code{stanfit}}}.
 objects are compared by the Bayes factor.
}
