% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_row_number_of_logical_vector.R
\name{the_row_number_of_logical_vector}
\alias{the_row_number_of_logical_vector}
\title{Extract the row number from a logical vector}
\usage{
the_row_number_of_logical_vector(vector.logical)
}
\arguments{
\item{vector.logical}{vector with logical component}
}
\value{
the row number of logical component
}
\description{
Extract the row number from a logical vector
}
\examples{


 a <-c(TRUE,FALSE,FALSE,TRUE)


the_row_number_of_logical_vector(a)

}
\author{
Issei Tsunoda
}
