% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_row_number_of_logical_vector.R
\name{TRUE.Counter.in.vector}
\alias{TRUE.Counter.in.vector}
\title{Count \code{TRUE} in a Vector whose components are all Logical \R objects}
\usage{
TRUE.Counter.in.vector(vector.logical)
}
\arguments{
\item{vector.logical}{vector with logical component}
}
\value{
A positive integer.
}
\description{
For the posterior predictive p value.
}
\examples{
#========================================================================================
#                         Revised 2019 oct. This is same as sum(), I did not know this
#========================================================================================

 a <-c(TRUE,FALSE,FALSE,TRUE,TRUE)

 TRUE.Counter.in.vector(a)

# Of course, it is:
#> Number.of.TRUE
#  3


sum(a) ==  TRUE.Counter.in.vector(a)

# I did not know this equality,... no longer this function is needed


}
