% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation.draw_srsc}
\alias{validation.draw_srsc}
\title{Draw Curves for validation dataset}
\usage{
validation.draw_srsc(validation.data, mesh.for.drawing.curve = 11111,
  upper_y = 1, DrawFROCcurve = TRUE)
}
\arguments{
\item{validation.data}{This is a return value of the function \code{validation.dataset_srsc}.}

\item{mesh.for.drawing.curve}{An integer indicating number of dots drawing the curves, default =10000.}

\item{upper_y}{This is a upper bound for the axis of the vertical coordinate of FROC curve.}

\item{DrawFROCcurve}{logical: TRUE or FALSE. Whether or not FROC curves are shown.}
}
\description{
drawing curves.
}
\examples{

\donttest{

    datasets <- validation.dataset_srsc()

    validation.draw_srsc(datasets)

}# dottest
}
